% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde_execution_functions.R
\name{adaptive_bw.mc}
\alias{adaptive_bw.mc}
\title{Function to calculate adaptative bandwidth according to Abramson’s smoothing regimen (multicore version)}
\usage{
adaptive_bw.mc(
  grid,
  events,
  lines,
  bw,
  trim_bw,
  method,
  kernel_name,
  max_depth,
  tol,
  digits,
  sparse,
  verbose
)
}
\arguments{
\item{grid}{a spatial grid to split the data within}

\item{events}{A spatialPointsDataFrame of the events points}

\item{lines}{A SpatialLinesDataFrame representing the network}

\item{bw}{the fixed kernel bandwidth}

\item{trim_bw}{the maximum size of local bandiwidths}

\item{method}{The method to use when calculating the NKDE}

\item{kernel_name}{The name of the kernel to use}

\item{max_depth}{the maximum recursion depth}

\item{tol}{a float indicating the spatial tolerance when snapping events on
lines}

\item{digits}{the number of digits to keep}

\item{sparse}{a boolean indicating if sparse matrix should be used}

\item{verbose}{a boolean indicating if update messages should be printed}
}
\value{
a vector with the local bandwidth
}
\description{
Function to calculate adaptative bandwidth according to Abramson’s smoothing regimen (multicore version)
}
\examples{
#This is an internal function, no example provided
}
