% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_kde.R
\name{adaptive_bw_1d}
\alias{adaptive_bw_1d}
\title{Adaptive bw in one dimension}
\usage{
adaptive_bw_1d(events, w, bw, kernel_name)
}
\arguments{
\item{events}{A numeric vector representing the moments of occurrence of events}

\item{w}{The weight of the events}

\item{bw}{A float, the bandiwdth to use}

\item{kernel_name}{The name of the kernel to use}

\item{samples}{A numeric vector representing the moments to sample}
}
\description{
Calculate adaptive bandwidths in one dimension
}
\keyword{internal}
