% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{add_center_lines}
\alias{add_center_lines}
\title{Add center vertex to lines}
\usage{
add_center_lines(lines)
}
\arguments{
\item{lines}{The feature collection of linestrings to use}
}
\value{
A feature collection of points
}
\description{
Add to each feature of a feature collection of lines an additional vertex at its
center.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
