% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{add_vertices}
\alias{add_vertices}
\title{Add vertices (SpatialPoints) to a single line (SpatialLines), may fail
if the lines geometries are self intersecting}
\usage{
add_vertices(line, points, i)
}
\arguments{
\item{line}{The SpatialLine to modify}

\item{points}{The SpatialPoints to add to as vertex to the lines}

\item{i}{The index of the line (for recombining after all lines)}
}
\value{
An object of class Lines (package sp)
}
\description{
Add vertices (SpatialPoints) to a single line (SpatialLines), may fail
if the lines geometries are self intersecting
}
\examples{
#This is an internal function, no example provided
}
