% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{bike_accidents}
\alias{bike_accidents}
\title{Road accidents including a bicyle in Montreal in 2016}
\format{
A sf object with 347 rows and 4 variables
\describe{
  \item{NB_VICTIME}{the number of victims}
  \item{AN}{the year of the accident}
  \item{Date}{the date of the accident (yyyy/mm/dd)}
  \item{geom}{the geometry (points)}
}
}
\source{
<https://donnees.montreal.ca/ville-de-montreal/collisions-routieres>
}
\usage{
bike_accidents
}
\description{
A feature collection (sf object) representing road accidents including a
cyclist in Montreal in 2016. The EPSG is 3797, and the data comes from the
Montreal OpenData website. It is only a small subset in central districts used to demonstrate the
main functions of spNetwork.
}
\keyword{datasets}
