% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{build_graph}
\alias{build_graph}
\title{Network generation}
\usage{
build_graph(lines, digits, line_weight, attrs = FALSE)
}
\arguments{
\item{lines}{A SpatialLinesDataFrame}

\item{digits}{The number of digits to keep from the coordinates}

\item{line_weight}{The name of a field that represent the cost to use a line}

\item{attrs}{A Boolean indicating if the original lines attributes
must be added to the graph lines}
}
\value{
A list containing the following elements:
\itemize{
        \item graph: an igraph object that preserves the original lines
        characteristics
        \item linelist: the dataframe used to build the graph
        \item lines: the original SpatialLinesDataFrame
        \item spvertices: a SpatialPointsDataFrame representing the vertices
        of the graph
        \item digits : the number of digits kept for the coordinates
}
}
\description{
Generate an igraph object from a SpatialLinesDataFrame.
}
\examples{
networkgpkg <- system.file("extdata", "networks.gpkg", package = "spNetwork", mustWork = TRUE)
mtl_network <- rgdal::readOGR(networkgpkg,layer="mtl_network", verbose=FALSE)
mtl_network$length <- rgeos::gLength(mtl_network, byid = TRUE)
graph_result <- build_graph(mtl_network, 2, "length", attrs = TRUE)
}
