% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{build_graph_directed}
\alias{build_graph_directed}
\title{Directed network generation}
\usage{
build_graph_directed(lines, digits, line_weight, direction, attrs = FALSE)
}
\arguments{
\item{lines}{A SpatialLinesDataFrame}

\item{digits}{The number of digits to keep from the coordinates}

\item{line_weight}{The name of a field that represent the cost to use a line}

\item{direction}{Indicate a field giving informations about authorized
traveling direction on lines. if NULL, then all lines can be used in both
directions. Must be the name of a column otherwise. The values of the
column must be "FT" (From - To), "TF" (To - From) or "Both".}

\item{attrs}{A boolean indicating if the original lines attributes}
}
\value{
A list containing the following elements:
\itemize{
        \item graph: an igraph object that preserves the original lines
        characteristics
        \item linelist: the dataframe used to build the graph
        \item lines: the original SpatialLinesDataFrame
        \item spvertices: a SpatialPointsDataFrame representing the vertices
        of the graph
        \item digits : the number of digits kept for the coordinates
}
}
\description{
Generate a directed igraph object from a SpatialLinesDataFrame.
}
\examples{
networkgpkg <- system.file("extdata", "networks.gpkg", package = "spNetwork", mustWork = TRUE)
mtl_network <- rgdal::readOGR(networkgpkg,layer="mtl_network", verbose=FALSE)
mtl_network$length <- rgeos::gLength(mtl_network, byid = TRUE)
mtl_network$direction <- "Both"
mtl_network[6, "direction"] <- "TF"
mtl_network_directed <- lines_direction(mtl_network, "direction")
graph_result <- build_graph_directed(lines = mtl_network_directed,
        digits = 2,
        line_weight = "length",
        direction = "direction",
        attrs = TRUE)
}
