% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{build_grid}
\alias{build_grid}
\title{Generate a grid of a specified shape in the bbox of a Spatial object}
\usage{
build_grid(grid_shape, spatial)
}
\arguments{
\item{grid_shape}{A numeric vector of length 2 indicating the number
of rows and the numbers of columns of the grid}

\item{spatial}{A list of SpatialDataFrames objects (package sp)}
}
\value{
a SpatialPolygonsDataFrame representing the grid
}
\description{
Generate a grid of a specified shape in the bbox of a Spatial object
}
\examples{
#This is an internal function, no example provided
}
