% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indexing_sf.R
\name{build_quadtree}
\alias{build_quadtree}
\title{Build a quadtree}
\usage{
build_quadtree(data)
}
\arguments{
\item{data}{a feature collection of linestrings or a feature collection of points}
}
\value{
a spatial_index object (pointer to a c++ instance)
}
\description{
Generate a quadtree object from package SearchTrees, useful to speed up
spatial requesting (INTERNAL).
}
\examples{
data(mtl_network)
tree <- build_quadtree(mtl_network)
buff <- sf::st_buffer(mtl_network[55,], 50)
selection <- spatial_request(buff, tree, mtl_network)
}
