% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{calc_gamma}
\alias{calc_gamma}
\title{Function calculating the gamma parameter in Abramson’s smoothing regimen}
\usage{
calc_gamma(k)
}
\arguments{
\item{k}{a vector of numeric values (the estimated kernel densities)}
}
\value{
the gamma parameter in Abramson’s smoothing regimen
}
\description{
Function calculating the gamma parameter in Abramson’s smoothing regimen
}
\examples{
#This is an internal function, no example provided
}
