% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde.R
\name{calc_NKDE}
\alias{calc_NKDE}
\title{Calculate the kernel density of a set of vertices in a graph}
\usage{
calc_NKDE(
  graph,
  origins,
  destinations,
  range,
  kernel = "quartic",
  weights = NULL
)
}
\arguments{
\item{graph}{The graph to use to calculate the densities}

\item{origins}{The indices of the vertices of the graph for which the
densities will be calculated}

\item{destinations}{The indices of the vertices to use as events for
the kernel density calculation}

\item{range}{The range value of the kernel to use}

\item{kernel}{The name of the kernel function to use (must be one of
quartic, gaussian or epanechnikov)}

\item{weights}{A vector of weights for the destination. Default is NULL
if NULL, all the destinations have a weight of 1.}
}
\value{
Vector of kernel densities (one for each origin)
}
\description{
Calculate the kernel density of a set of vertices in a graph
}
\examples{
#This is an internal function, no example provided
}
