% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde_execution_functions.R
\name{clean_events}
\alias{clean_events}
\title{Function to avoid having events at the same location}
\usage{
clean_events(events, digits = 5)
}
\arguments{
\item{events}{the SpatialPointsDataFrame to contract (must have a weight column)}

\item{digits}{the number of digits to keep}
}
\value{
a new SpatialPointsDataFrame
}
\description{
Function to avoid having events at the same location
}
\examples{
#This is an internal function, no example provided
}
