% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indexing.R
\name{closest_points}
\alias{closest_points}
\title{build a quad tree index and solve the nearest neighbour problem for two
SpatialPointsDataFrame}
\usage{
closest_points(points, features)
}
\arguments{
\item{points}{a SpatialPointsDataFrame}

\item{features}{a SpatialPointsDataFrame}
}
\value{
for each point, the index of the nearest line (row-wise)
}
\description{
build a quad tree index and solve the nearest neighbour problem for two
SpatialPointsDataFrame
}
\examples{
#This is an internal function, no example provided
}
