% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indexing.R
\name{clostest_features}
\alias{clostest_features}
\title{build a quad tree index and solve the nearest neighbour problenm for points
to lines}
\usage{
clostest_features(points, lines)
}
\arguments{
\item{points}{a SpatialPointsDataFrame}

\item{lines}{a SpatialLinesDataFrame}
}
\value{
for each point, the index of the nearest line (row-wise)
}
\description{
build a quad tree index and solve the nearest neighbour problenm for points
to lines
}
\examples{
#This is an internal function, no example provided
}
