% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border_correction.R
\name{correction_factor_time}
\alias{correction_factor_time}
\title{Time extent correction for NKDE}
\usage{
correction_factor_time(events_time, samples_time, bws_time, kernel_name)
}
\arguments{
\item{events_time}{A numeric vector representing when the events occurred}

\item{samples_time}{A numeric vector representing when the densities will
be sampled}

\item{bws_time}{A numeric vector with the temporal bandwidths}

\item{kernel_name}{The name of the kernel to use}
}
\value{
A numeric vector with the correction factor values for each event
}
\description{
Function to calculate the time extent correction factor in tnkde.
}
\examples{
#no example provided, this is an internal function
}
\keyword{internal}
