% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border_correction_sf.R
\name{corrfactor_simple}
\alias{corrfactor_simple}
\title{Simple NKDE border correction}
\usage{
corrfactor_simple(graph, events, edges, bws)
}
\arguments{
\item{graph}{The graph (igraph) used to calculate distances between nodes}

\item{events}{A feature collection of points representing the events}

\item{edges}{A feature collection of lines representing the edges of the graph}

\item{bws}{A vector of the bandwidths for each event}
}
\value{
A numeric vector with the correction factor values
}
\description{
A function to calculate the Diggle correction factor with the
simple NKDE.
}
\keyword{internal}
