% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{cosine_kernel}
\alias{cosine_kernel}
\title{cosine kernel}
\usage{
cosine_kernel(d, bw)
}
\arguments{
\item{d}{the distance from the event}

\item{bw}{the bandwidth used for the kernel}
}
\value{
the estimated density
}
\description{
cosine kernel
}
\examples{
#This is an internal function, no example provided
}
