% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosine_kernel_cpp}
\alias{cosine_kernel_cpp}
\title{c++ cosine kernel}
\usage{
cosine_kernel_cpp(d, bw)
}
\arguments{
\item{d}{a vector of distances for which the density must be calculated}

\item{bw}{a double representing the size of the kernel bandwidth}
}
\description{
c++ cosine kernel
}
\keyword{internal}
