% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosine_kernelos}
\alias{cosine_kernelos}
\title{c++ cosine kernel for one distance}
\usage{
cosine_kernelos(d, bw)
}
\arguments{
\item{d}{a double, the distances for which the density must be calculated}

\item{bw}{a double representing the size of the kernel bandwidth}
}
\description{
c++ cosine kernel for one distance
}
\keyword{internal}
