% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_functions.R
\name{cross_gfun}
\alias{cross_gfun}
\title{Base cross g-function}
\usage{
cross_gfun(dist_mat, start, end, step, width, Lt, na, nb, wa, wb)
}
\arguments{
\item{dist_mat}{A matrix with the distances between points}

\item{start}{A float, the start value for evaluating the g-function}

\item{end}{A float, the last value for evaluating the g-function}

\item{step}{A float, the jump between two evaluations of the k-function}

\item{width}{The width of each donut}

\item{Lt}{The total lenght of the network}

\item{na}{The number of points in set A}

\item{nb}{The number of points in set B}

\item{wa}{The weight of the points in set A (coincident points)}

\item{wb}{The weight of the points in set B (coincident points)}
}
\value{
A numeric vector with the values
}
\description{
Calculate the base cross g-function.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
