% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cross_kfunc_cpp}
\alias{cross_kfunc_cpp}
\title{c++ cross k function}
\usage{
cross_kfunc_cpp(dist_mat, start, end, step, Lt, na, nb, wa, wb)
}
\arguments{
\item{dist_mat}{A square matrix with the distances between points}

\item{start}{A float, the start value for evaluating the k-function}

\item{end}{A float, the last value for evaluating the k-function}

\item{step}{A float, the jump between two evaluations of the k-function}

\item{Lt}{The total length of the network}

\item{na}{The number of points in set A}

\item{nb}{The number of points in set B}

\item{wa}{The weight of the points in set A (coincident points)}

\item{wb}{The weight of the points in set B (coincident points)}
}
\description{
c++ cross k function
}
