% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions_sf.R
\name{direct_lines}
\alias{direct_lines}
\title{Make a network directed}
\usage{
direct_lines(lines, direction)
}
\arguments{
\item{lines}{The original feature collection of linestrings}

\item{direction}{A vector of integers. 0 indicates a bidirectional line and 1
an unidirectional line}
}
\value{
A feature collection of linestrings with some lines duplicated according to
direction
}
\description{
Function to create complementary lines for a directed network.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
