% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{discontinuous_nkde2}
\alias{discontinuous_nkde2}
\title{Function to perform the discontinuous nkde}
\usage{
discontinuous_nkde2(
  edge_list,
  neighbour_list,
  v_events,
  weights,
  samples,
  bw,
  kernel_func,
  nodes,
  linelist,
  verbose
)
}
\arguments{
\item{edge_list}{a list of the edges, accessible with the names of the
two nodes composing the edge}

\item{neighbour_list}{a list indicating for each node its neighbours
must indicate for each event its corresponding node}

\item{v_events}{a numeric vector of the event nodes}

\item{weights}{a numeric vector of the weights of the events}

\item{samples}{a SpatialPointsDataFrame of the sampling points}

\item{bw}{the kernel bandwidth}

\item{kernel_func}{a function obtained with the function select_kernel}

\item{nodes}{a SpatialPointsDataFrame representing the nodes of the network}

\item{linelist}{the linelist of the network (igraph)}

\item{verbose}{a boolean indicating if messages must be displayed}
}
\value{
a dataframe with two columns. sum_k is the sum for each sample point
 of the kernel values. n is the number of events influencing each sample
point
}
\description{
Function to perform the discontinuous nkde
}
\examples{
#This is an internal function, no example provided
}
