% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discontinuousWorker_int}
\alias{discontinuousWorker_int}
\title{The worker function to calculate discontinuous NKDE (with ARMADILLO and Integer matrix)}
\arguments{
\item{kernel_func}{a cpp pointer function (selected with the kernel name)}

\item{edge_mat}{matrix, to find the id of each edge given two neighbours}

\item{neighbour_list}{a List, giving for each node an IntegerVector with
its neighbours}

\item{v}{the actual node to consider for the recursion (int)}

\item{bw}{the kernel bandiwdth}

\item{line_weights}{a vector with the length of the edges}

\item{samples_edgeid}{a vector associating each sample to an edge}

\item{samples_x}{a vector with x coordinates of each sample}

\item{samples_ya}{vector with y coordinates of each sample}

\item{nodes_x}{a vector with x coordinates of each node}

\item{nodes_y}{a vector with y coordinates of each node}

\item{depth}{the actual recursion depth}

\item{max_depth}{the maximum recursion depth}
}
\value{
a vector with the kernel values calculated for each samples from
the first node given
}
\description{
The worker function to calculate discontinuous NKDE (with ARMADILLO and Integer matrix)
}
