% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discontinuousWorker_sparse}
\alias{discontinuousWorker_sparse}
\title{The worker function to calculate discontinuous NKDE (with ARMADILLO and sparse matrix)}
\arguments{
\item{kernel_func}{a cpp pointer function (selected with the kernel name)}

\item{edge_mat}{matrix, to find the id of each edge given two neighbours}

\item{neighbour_list}{a List, giving for each node an IntegerVector with
its neighbours}

\item{v}{the actual node to consider for the recursion (int)}

\item{bw}{the kernel bandiwdth}

\item{line_weights}{a vector with the length of the edges}

\item{samples_edgeid}{a vector associating each sample to an edge}

\item{samples_coords}{a matrix with the X and Y coordinates of the samples}

\item{nodes_coords}{a matrix with the X and Y coordinates of the nodes}

\item{depth}{the actual recursion depth}

\item{max_depth}{the maximum recursion depth}
}
\value{
a vector with the kernel values calculated for each samples from
the first node given
}
\description{
The worker function to calculate discontinuous NKDE (with ARMADILLO and sparse matrix)
}
\keyword{internal}
