% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde.R
\name{epanechnikov_kernel}
\alias{epanechnikov_kernel}
\title{calcualte the density at a point with a epanechnikov function given
the distances of the points around, their weights and the kernel range}
\usage{
epanechnikov_kernel(d, w, r)
}
\arguments{
\item{d}{A vector of distances}

\item{w}{A vector of weights}

\item{r}{A float representing the range}
}
\value{
The kernel density
}
\description{
calcualte the density at a point with a epanechnikov function given
the distances of the points around, their weights and the kernel range
}
\examples{
#This is an internal function, no example provided
}
