% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{epanechnikov_kernelos}
\alias{epanechnikov_kernelos}
\title{c++ epanechnikov kernel for one distance}
\usage{
epanechnikov_kernelos(d, bw)
}
\arguments{
\item{d}{a double, the distances for which the density must be calculated}

\item{bw}{a double representing the size of the kernel bandwidth}
}
\description{
c++ epanechnikov kernel for one distance
}
\keyword{internal}
