% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{esc_kernel_loo_nkde}
\alias{esc_kernel_loo_nkde}
\title{The worker function to calculate continuous TNKDE likelihood cv}
\arguments{
\item{kernel_func}{a cpp pointer function (selected with the kernel name)}

\item{edge_mat}{matrix, to find the id of each edge given two neighbours.}

\item{events}{a NumericVector indicating the nodes in graph beeing events}

\item{neighbour_list}{a List, giving for each node an IntegerVector with
its neighbours}

\item{v}{the actual node to consider (int)}

\item{bws_net}{an arma::vec with the network bandwidths to consider}

\item{line_weights}{a vector with the length of the edges}

\item{depth}{the actual recursion depth}

\item{max_depth}{the maximum recursion depth}
}
\value{
a cube with the impact of the event v on each other event for
each pair of bandwidths (cube(bws_net, bws_time, events))
}
\description{
The worker function to calculate continuous TNKDE likelihood cv
}
