% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{esd_kernel2}
\alias{esd_kernel2}
\title{Worker function for the discontinuous nkde}
\usage{
esd_kernel2(
  y,
  edge_list,
  neighbour_list,
  samples,
  bw,
  kernel_func,
  nodes,
  lines_weight,
  linelist,
  verbose
)
}
\arguments{
\item{y}{the vertex index of the actual event}

\item{edge_list}{a list of the edges, accessible with the names of the
two nodes composing the edge}

\item{neighbour_list}{a list indicating for each node its neighbours
must indicate for each event its corresponding node}

\item{samples}{a SpatialPointsDataFrame of the sampling points}

\item{bw}{the kernel bandwidth}

\item{kernel_func}{a function obtained with the function select_kernel}

\item{nodes}{a SpatialPointsDataFrame representing the nodes of the network}

\item{lines_weight}{a numeric vector with the weight of the lines}

\item{linelist}{the linelist of the network (igraph)}

\item{verbose}{a boolean indicating if messages must be displayed}
}
\description{
Worker function for the discontinuous nkde
}
\examples{
#This is an internal function, no example provided
}
