% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{esd_kernel_loo}
\alias{esd_kernel_loo}
\title{The function to calculate discontinuous NKDE likelihood cv}
\arguments{
\item{kernel_func}{a cpp pointer function (selected with the kernel name)}

\item{samples_k}{a numeric vector of the actual kernel values, updated at
each recursion}

\item{neighbour_list}{a List, giving for each node an IntegerVector with
its neighbours}

\item{edge_mat}{matrix, to find the id of each edge given two neighbours.}

\item{v}{the actual node to consider for the recursion (int)}

\item{bw}{the kernel bandwidth}

\item{line_weights}{a vector with the length of the edges}

\item{depth}{the actual recursion depth}

\item{max_depth}{the maximum recursion depth}
}
\value{
a vector with the kernel values calculated for each samples from
the first node given

a vector with the kernel values calculated for each samples from
the first node given
}
\description{
The function to calculate discontinuous NKDE likelihood cv

The function to calculate simple NKDE likelihood cv
}
\keyword{internal}
