% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ess_kernel_loo_nkde}
\alias{ess_kernel_loo_nkde}
\title{The worker function to calculate simple NKDE likelihood cv}
\arguments{
\item{kernel_func}{a cpp pointer function (selected with the kernel name)}

\item{edge_mat}{matrix, to find the id of each edge given two neighbours.}

\item{events}{a NumericVector indicating the nodes in the graph being events}

\item{neighbour_list}{a List, giving for each node an IntegerVector with
its neighbours}

\item{v}{the actual node to consider (int)}

\item{bws_net}{an arma::vec with the network bandwidths to consider}

\item{line_weights}{a vector with the length of the edges}

\item{depth}{the actual recursion depth}

\item{max_depth}{the maximum recursion depth}
}
\value{
a matrix with the impact of the event v on each other events for
each pair of bandwidths (mat(event, bws_net))
}
\description{
The worker function to calculate simple NKDE likelihood cv
}
\keyword{internal}
