% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_neighbouring.R
\name{exe_line_center_listw}
\alias{exe_line_center_listw}
\title{Worker function for the line_center_listw_gridded.mc function.}
\usage{
exe_line_center_listw(
  i,
  lines,
  centers,
  grid,
  maxdistance,
  digits,
  matrice_type,
  vdist_func,
  direction
)
}
\arguments{
\item{i}{The row index of the spatiale grid to use}

\item{lines}{The SpatialLinesDataFrame to use as a network}

\item{centers}{The centers of the lines (SpatialPointsDataFrame)}

\item{grid}{The SpatialPolygons object reprsenting the grid for splitted
calculation}

\item{maxdistance}{The maximum distance between two observation to
considere them as neighbours.}

\item{digits}{the number of digits to keep in the spatial coordinates (
simplification used to reduce risk of topological error)}

\item{matrice_type}{The type of the weighting scheme. Can be 'B' for Binary,
'W' for row weighted, see the documentation of spdep::nb2listw for details}

\item{vdist_func}{Indicates the function to use to convert the distance
between observation in spatial weights. Can be 'identity', 'inverse',
'squared inverse' or a function with one parameter x that will be
vectorized internally}

\item{direction}{indicate a field giving informations about authorized
traveling direction on lines. if NULL, then all lines can be used in both
directions. Must be the name of a column otherwise. The values of the
column must be "FT" (From - To), "TF" (To - From) or "Both".}
}
\value{
A list, containing a neighbours list and a weights list
}
\description{
Worker function for the line_center_listw_gridded.mc function.
}
\examples{
#This is an internal function, no example provided
}
