% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_neighbouring.R
\name{exe_line_ext_listw}
\alias{exe_line_ext_listw}
\title{Worker function for the line_ext_listw_gridded.mc function.}
\usage{
exe_line_ext_listw(
  i,
  lines,
  grid,
  maxdistance,
  digits,
  matrice_type,
  vdist_func,
  mindist,
  direction
)
}
\arguments{
\item{i}{The row index of the spatiale grid to use}

\item{lines}{The SpatialLinesDataFrame to use as a network}

\item{grid}{The SpatialPolygons object reprsenting the grid for splitted
calculation}

\item{maxdistance}{The maximum distance between two observation to
considere them as neighbours.}

\item{digits}{the number of digits to keep in the spatial coordinates (
simplification used to reduce risk of topological error)}

\item{matrice_type}{The type of the weighting scheme. Can be 'B' for Binary,
'W' for row weighted, see the documentation of spdep::nb2listw for details}

\item{vdist_func}{A function to convert distance in spatial weights}

\item{mindist}{Indicates the minimum value to replace 0 with. Two lines that
share a vertex have a 0 distance between them, wich could create some
trouble in the weightings of the neighbouring object.}

\item{direction}{indicate a field giving informations about authorized
traveling direction on lines. if NULL, then all lines can be used in both
directions. Must be the name of a column otherwise. The values of the
column must be "FT" (From - To), "TF" (To - From) or "Both".}
}
\value{
A list, containing a neighbours list and a weights list
}
\description{
Worker function for the line_ext_listw_gridded.mc function.
}
\examples{
#This is an internal function, no example provided
}
