% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_nkde.R
\name{exe_nkde}
\alias{exe_nkde}
\title{Worker function for the nkde_grided.mc function.}
\usage{
exe_nkde(
  i,
  grid,
  lines,
  lixels,
  points,
  kernel,
  kernel_range,
  snap_dist,
  digits,
  tol
)
}
\arguments{
\item{i}{The row index of the spatiale grid to use}

\item{grid}{The SpatialPolygons object reprsenting the grid for splitted
calculation}

\item{lines}{The SpatialLinesDataFrame to use as a network}

\item{points}{The events to use in the kernel density estimation}

\item{kernel}{The name of the kernel function to use (must be one of
quartic, gaussian or epanechnikov). Default is Quartic}

\item{kernel_range}{The range of the kernel function}

\item{snap_dist}{The maximum distance snapping between points and lines}

\item{digits}{The number of digits to keep in the coordinates of the
geometries}

\item{tol}{The tolerence for topological operations}

\item{weights}{A vector of weights for the events. Default is NULL
if NULL, all the events have a weight of 1.
track the progressing of the calculation.}
}
\value{
A SpatialLinesDataFrame of the lines and their kernel densities
}
\description{
Worker function for the nkde_grided.mc function.
}
\examples{
#This is an internal function, no example provided
}
