% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{g_nt_func_cpp}
\alias{g_nt_func_cpp}
\title{c++ g space-time function}
\usage{
g_nt_func_cpp(
  dist_mat_net,
  dist_mat_time,
  start_net,
  end_net,
  step_net,
  width_net,
  start_time,
  end_time,
  step_time,
  width_time,
  Lt,
  Tt,
  n,
  w
)
}
\arguments{
\item{dist_mat_time}{A square matrix with the distances between points in time}

\item{start_net}{A float, the start value for evaluating the g-function on the network}

\item{end_net}{A float, the last value for evaluating the g-function on the network}

\item{step_net}{A float, the jump between two evaluations of the g-function on the network}

\item{width_net}{The width of each donut on the network}

\item{start_time}{A float, the start value for evaluating the g-function in time}

\item{end_time}{A float, the last value for evaluating the g-function in time}

\item{step_time}{A float, the jump between two evaluations of the g-function in time}

\item{width_time}{The width of each donut in time}

\item{Lt}{The total length of the network}

\item{n}{The number of points}

\item{w}{The weight of the points (coincident points)}

\item{dist_mat}{A square matrix with the distances between points on the network}
}
\description{
c++ g space-time function
}
\keyword{internal}
