% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussian_kernel_scaledos}
\alias{gaussian_kernel_scaledos}
\title{c++ scaled gaussian kernel for one distance}
\usage{
gaussian_kernel_scaledos(d, bw)
}
\arguments{
\item{d}{a double, the distances for which the density must be calculated}

\item{bw}{a double representing the size of the kernel bandwidth}
}
\description{
c++ scaled gaussian kernel for one distance
}
\keyword{internal}
