% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_loo_values_simple}
\alias{get_loo_values_simple}
\title{The exported function to calculate simple NKDE likelihood cv}
\usage{
get_loo_values_simple(
  neighbour_list,
  samples,
  sweights,
  events,
  weights,
  bws,
  kernel_name,
  line_list,
  max_depth
)
}
\arguments{
\item{neighbour_list}{a List, giving for each node an IntegerVector with
its neighbours}

\item{samples}{a vector with the node index of the samples}

\item{sweights}{a vector with the weight of each sample}

\item{events}{a vector with the node index of the events}

\item{weights}{a vector with the weights of the events}

\item{bws}{a vector with the bandwidths}

\item{kernel_name}{the name of the kernel function to use}

\item{line_list}{a dataframe representing the edges of the network}

\item{max_depth}{the maximum recursion depth}
}
\value{
a vector with the kernel values calculated for each event
}
\description{
The exported function to calculate simple NKDE likelihood cv
}
\keyword{interal}
