% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_functions.R
\name{gfunc}
\alias{gfunc}
\title{Base g-function}
\usage{
gfunc(dist_mat, start, end, step, width, Lt, n, w)
}
\arguments{
\item{dist_mat}{A square matrix with the distances between points}

\item{start}{A float, the start value for evaluating the g-function}

\item{end}{A float, the last value for evaluating the g-function}

\item{step}{A float, the jump between two evaluations of the k-function}

\item{width}{The width of each donut}

\item{Lt}{The total length of the network}

\item{n}{The number of points}

\item{w}{The weight of the points (coincident points)}
}
\value{
A numeric vector with the values
}
\description{
Calculate the base g-function.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
