% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gfunc_cpp2}
\alias{gfunc_cpp2}
\title{c++ g function}
\usage{
gfunc_cpp2(dist_mat, start, end, step, width, Lt, n, wc, wr)
}
\arguments{
\item{dist_mat}{A square matrix with the distances between points}

\item{start}{A float, the start value for evaluating the g-function}

\item{end}{A float, the last value for evaluating the g-function}

\item{step}{A float, the jump between two evaluations of the k-function}

\item{width}{The width of each donut}

\item{Lt}{The total length of the network}

\item{n}{The number of points}

\item{wc}{The weight of the points represented by the columns (destinations)}

\item{wr}{The weight of the points represented by the rows (origins)}
}
\value{
A numeric vector with the values of the g function evaluated at the required distances
}
\description{
c++ g function (INTERNAL)
}
