% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{gm_mean}
\alias{gm_mean}
\title{Function calculating the geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector of numeric values}

\item{na.rm}{a boolean indicating if we filter the NA values}
}
\value{
the geometric mean of x
}
\description{
Function calculating the geometric mean
}
\examples{
#This is an internal function, no example provided
}
