% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{heal_edges}
\alias{heal_edges}
\title{Heal edges}
\usage{
heal_edges(lines, digits = 3, verbose = TRUE)
}
\arguments{
\item{lines}{A feature collection of linestrings}

\item{digits}{An integer indicating the number of digits to keep in coordinates}

\item{verbose}{A boolean indicating if a progress bar should be displayed}
}
\value{
A feature collection of linestrings with the eventually merged geometries. Note
that if lines are merged, only the attributes of the first line are preserved
}
\description{
Merge Lines if they form a longer linestring without external intersections (experimental)
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
