% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde_execution_functions_sf.R
\name{is_projected}
\alias{is_projected}
\title{Projection test}
\usage{
is_projected(obj)
}
\arguments{
\item{obj}{A feature collection}
}
\value{
A boolean
}
\description{
Check if a feature collection is in a projected CRS
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
