% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{k_nt_func_cpp}
\alias{k_nt_func_cpp}
\alias{k_g_nt_func_cpp2}
\title{c++ k space-time function}
\usage{
k_nt_func_cpp(
  dist_mat_net,
  dist_mat_time,
  start_net,
  end_net,
  step_net,
  start_time,
  end_time,
  step_time,
  Lt,
  Tt,
  n,
  w
)

k_g_nt_func_cpp2(
  dist_mat_net,
  dist_mat_time,
  start_net,
  end_net,
  step_net,
  start_time,
  end_time,
  step_time,
  width_net,
  width_time,
  Lt,
  Tt,
  n,
  wc,
  wr,
  cross = TRUE
)
}
\arguments{
\item{dist_mat_net}{A square matrix with the distances between points (network)}

\item{dist_mat_time}{A square matrix with the distances between points (time)}

\item{start_net}{A float, the start value for evaluating the k-function (network)}

\item{end_net}{A float, the last value for evaluating the k-function (network)}

\item{step_net}{A float, the jump between two evaluations of the k-function (network)}

\item{start_time}{A float, the start value for evaluating the k-function (time)}

\item{end_time}{A float, the last value for evaluating the k-function (time)}

\item{step_time}{A float, the jump between two evaluations of the k-function (time)}

\item{Lt}{The total length of the network}

\item{Tt}{The total duration of study area}

\item{n}{The number of points}

\item{w}{The weight of the points (coincident points)}

\item{cross}{a boolean indicating of we are calculating a cross k or g function}
}
\description{
c++ k space-time function

c++ k space-time function
}
\keyword{internal}
