% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_functions.R
\name{kfunc}
\alias{kfunc}
\title{Base k-function}
\usage{
kfunc(dist_mat, start, end, step, Lt, n, w)
}
\arguments{
\item{dist_mat}{A square matrix with the distances between points}

\item{start}{A float, the start value for evaluating the k-function}

\item{end}{A float, the last value for evaluating the k-function}

\item{step}{A float, the jump between two evaluations of the k-function}

\item{Lt}{The total lenght of the network}

\item{n}{The number of points}

\item{w}{The weight of the points (coincident points)}
}
\value{
A numeric vector with the values
}
\description{
Calculate the base k-function.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
