% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kfunc_counting}
\alias{kfunc_counting}
\title{c++ k function counting worker}
\usage{
kfunc_counting(dist_mat, wc, wr, breaks, cross = FALSE)
}
\arguments{
\item{dist_mat}{A matrix with the distances between points}

\item{wc}{The weight of the points represented by the columns (destinations)}

\item{wr}{The weight of the points represented by the rows (origins)}

\item{breaks}{A numeric vector with the distance to consider}

\item{cross}{A boolean indicating if we are calculating a cross k function or not (default is FALSE)}
}
\value{
A numeric matrix with the countings of the k function evaluated at the required distances
}
\description{
c++ k function counting (INTERNAL)
}
\keyword{internal}
