% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kfunc_cpp}
\alias{kfunc_cpp}
\title{c++ k function}
\usage{
kfunc_cpp(dist_mat, start, end, step, Lt, n, w)
}
\arguments{
\item{dist_mat}{A square matrix with the distances between points}

\item{start}{A float, the start value for evaluating the k-function}

\item{end}{A float, the last value for evaluating the k-function}

\item{step}{A float, the jump between two evaluations of the k-function}

\item{Lt}{The total length of the network}

\item{n}{The number of points}

\item{w}{The weight of the points (coincident points)}
}
\value{
A numeric vector with the values of the k function evaluated at the required distances
}
\description{
c++ k function (INTERNAL)
}
