% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kgfunc_counting}
\alias{kgfunc_counting}
\title{c++ k and g function counting worker}
\usage{
kgfunc_counting(dist_mat, wc, wr, breaks, width, cross = FALSE)
}
\arguments{
\item{dist_mat}{A matrix with the distances between points}

\item{wc}{The weight of the points represented by the columns (destinations)}

\item{wr}{The weight of the points represented by the rows (origins)}

\item{breaks}{A numeric vector with the distance to consider}

\item{width}{The width of each donut}

\item{cross}{A boolean indicating if we are calculating a cross k function or not (default is FALSE)}
}
\value{
A list  of two numeric matrices with the values of the k and g function evaluated at the required distances
}
\description{
c++ k function counting (INTERNAL)
}
