% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kgfunc_time_counting}
\alias{kgfunc_time_counting}
\title{c++ k and g function counting worker}
\usage{
kgfunc_time_counting(
  dist_mat_net,
  dist_mat_time,
  wc,
  wr,
  breaks_net,
  breaks_time,
  width_net,
  width_time,
  cross = FALSE
)
}
\arguments{
\item{dist_mat_net}{A matrix with the distances between points on the network}

\item{dist_mat_time}{A matrix with the distances between points in time}

\item{wc}{The weight of the points represented by the columns (destinations)}

\item{wr}{The weight of the points represented by the rows (origins)}

\item{breaks_net}{A numeric vector with the distance to consider on network}

\item{breaks_time}{A numeric vector with the distance to consider in time}

\item{width_net}{The width of each donut for the network dimension}

\item{width_time}{The width of each donut for the time dimension}

\item{cross}{A boolean indicating if we are calculating a cross k function or not (default is FALSE)}
}
\value{
A list  of two numeric matrices with the values of the k and g function evaluated at the required distances
}
\description{
c++ k function counting (INTERNAL)
}
