% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{libraries_mtl}
\alias{libraries_mtl}
\title{Libraries}
\format{
A SpatialPointsDataFrame with 55 rows and 2 variables
}
\source{
\url{http://donnees.ville.montreal.qc.ca/}
}
\usage{
libraries_mtl
}
\description{
A SpatialPointsDataFrame representing some libraries in Montreal
}
\keyword{datasets}
