% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouring.R
\name{line_center_listw_gridded}
\alias{line_center_listw_gridded}
\title{generate listw object (spdep like) based on network distances}
\usage{
line_center_listw_gridded(
  lines,
  maxdistance,
  dist_func = "inverse",
  matrice_type = "B",
  grid_shape = c(2, 2),
  show_progress = TRUE,
  digits = 3
)
}
\arguments{
\item{lines}{A SpatialLinesDataFrame}

\item{maxdistance}{The maximum distance between two observation to
considere them as neighbours.}

\item{dist_func}{Indicates the function to use to convert the distance
between observation in spatial weights. Can be "identity", "inverse",
"squared inverse" or a function with one parameter x that will be
vectorized internally}

\item{matrice_type}{The type of the weighting scheme. Can be "B" for Binary,
"W" for row weighted, see the documentation of spdep::nb2listw for details}

\item{grid_shape}{A vector of length 2 indicating the shape of the grid to
use for splitting the dataset. Default is c(2,2)}

\item{show_progress}{A boolean indicating if a plot should be displayed to
track the progressing of the calculation.}

\item{digits}{the number of digits to keep in the spatial coordinates (
simplification used to reduce risk of topological error)}

\item{mindist}{Indicates the minimum value to replace 0 with. Two lines that
share a vertex have a 0 distance between them, wich could create some
trouble in the weightings of the neighbouring object.}
}
\value{
A listw object (spdep like)
}
\description{
The distances are calculated from the centroid of each line to the centroid
of each other lines.
}
\examples{
data(mtl_network)
listw <- line_center_listw_gridded(mtl_network,maxdistance=800,
        dist_func = "squared inverse", matrice_type="B",
        grid_shape = c(5,5))
}
