% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{lines_center}
\alias{lines_center}
\title{Center points of lines}
\usage{
lines_center(lines, verbose = FALSE)
}
\arguments{
\item{lines}{The SpatialLinesDataframe to use}

\item{verbose}{A Boolean indicating if a progressbar should be displayed}
}
\value{
An object of class SpatialPointsDataFrame (package sp)
}
\description{
Generate a SpatialPointsDataFrame with line center points. The points are
located at center of the line based on the length of the line.
}
\examples{
data('mtl_network')
centers <- lines_center(mtl_network)
}
