% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{lines_coordinates_as_list}
\alias{lines_coordinates_as_list}
\title{Lines coordinates as list}
\usage{
lines_coordinates_as_list(lines)
}
\arguments{
\item{lines}{A sf object with linestring type geometries}
}
\value{
A list of matrices
}
\description{
A function to get the coordinates of some lines as a list of matrices
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
