% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{lines_points_along}
\alias{lines_points_along}
\title{Generate points along the lines of a SpatialLinesDataFrame}
\usage{
lines_points_along(lines, dist)
}
\arguments{
\item{lines}{The SpatialLinesDataframe to use}

\item{dist}{the distance between the points along the lines}
}
\value{
An object of class SpatialLinesDataframe (package sp)
}
\description{
Generate points along the lines of a SpatialLinesDataFrame
}
\examples{
data(mtl_network)
new_pts <- lines_points_along(mtl_network,50)
}
