% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{list_coordinates_as_lines}
\alias{list_coordinates_as_lines}
\title{List of coordinates as lines}
\usage{
list_coordinates_as_lines(coord_list, crs)
}
\arguments{
\item{lines}{A list of matrices}
}
\value{
A sf object with linestring type geometries
}
\description{
A function to convert a list of matrices to as sf object with linestring geometry type
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
