% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{main_network_mtl}
\alias{main_network_mtl}
\title{Primary road network of Montreal}
\format{
A sf object with 2945 rows and 2 variables
\describe{
\item{TYPE}{the type of road}
\item{geom}{the geometry (linestrings)}
}
}
\source{
\url{https://donnees.montreal.ca/dataset/geobase}
}
\usage{
main_network_mtl
}
\description{
A feature collection (sf object) representing the primary road network of Montreal. The EPSG is 3797, and the data comes from the
Montreal OpenData website.
}
\keyword{datasets}
