% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mtl_libraries}
\alias{mtl_libraries}
\title{Libraries}
\format{
A SpatialPointsDataFrame with 55 rows and 2 variables
}
\source{
\url{http://donnees.ville.montreal.qc.ca/}
}
\usage{
mtl_libraries
}
\description{
A SpatialPointsDataFrame representing libraries in Montreal
}
\keyword{datasets}
