% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mtl_network}
\alias{mtl_network}
\title{An extract of the Montreal road network}
\format{
A SpatialLinesDataFrame with 2946 rows and 4 variables:
\describe{
  \item{IdRte}{a unique id for each road}
  \item{ClsRte}{the type of the road}
  \item{ClsFnction}{the function of the road}
  \item{nbAccident}{the number of
  bike accidents recorded on the road} ... }
}
\source{
\url{http://donnees.ville.montreal.qc.ca/}
}
\usage{
mtl_network
}
\description{
A SpatialLinesDataFrame representing a portion of the road network of
Montreal.
}
\keyword{datasets}
