% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mtl_theatres}
\alias{mtl_theatres}
\title{Theatres}
\format{
A SpatialPointsDataFrame with 54 rows and 2 variables
}
\source{
\url{http://donnees.ville.montreal.qc.ca/}
}
\usage{
mtl_theatres
}
\description{
A SpatialPointsDataFrame representing theatres in Montreal
}
\keyword{datasets}
