% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{nearest}
\alias{nearest}
\title{Nearest feature}
\usage{
nearest(x, y)
}
\arguments{
\item{x}{A SpatialDataFrame}

\item{y}{A SpatialDataFrame}
}
\value{
A numeric vector with the index of the nearest features
}
\description{
Find the nearest feature from set X in set Y
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
