% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{nearestPointOnLine}
\alias{nearestPointOnLine}
\title{Nearest point on Line}
\usage{
nearestPointOnLine(coordsLine, coordsPoint)
}
\arguments{
\item{coordsLine}{The coordinates of the line (matrix)}

\item{coordsPoint}{The coordinates of the point}
}
\value{
A numeric vector with the coordinates of the projected point
}
\description{
Find the nearest projected point on a LineString (from maptools)
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
