% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{nearestPointOnSegment}
\alias{nearestPointOnSegment}
\title{Nearest point on segment}
\usage{
nearestPointOnSegment(s, p)
}
\arguments{
\item{s}{The coordinates of the segment}

\item{p}{The coordinates of the point}
}
\value{
A numeric vector with the coordinates of the projected point
}
\description{
Find the nearest projected point on a segment (from maptools)
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
