% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_nkde.R
\name{nkde_grided.mc}
\alias{nkde_grided.mc}
\title{Gridded version of nkde with multicore support.}
\usage{
nkde_grided.mc(
  lines,
  points,
  snap_dist,
  lixel_length,
  kernel_range,
  kernel = "quartic",
  tol = 0.1,
  digits = 3,
  weights = NULL,
  grid_shape = c(2, 2),
  show_progress = TRUE
)
}
\arguments{
\item{lines}{The SpatialLinesDataFrame to use as a network}

\item{points}{The events to use in the kernel density estimation}

\item{snap_dist}{The maximum distance snapping between points and lines}

\item{lixel_length}{The expected length of a lixel}

\item{kernel_range}{The range of the kernel function}

\item{kernel}{The name of the kernel function to use (must be one of
quartic, gaussian or epanechnikov). Default is Quartic}

\item{tol}{The tolerence for topological operations}

\item{digits}{The number of digits to keep in the coordinates of the
geometries}

\item{weights}{A vector of weights for the events. Default is NULL
if NULL, all the events have a weight of 1.}

\item{grid_shape}{A vector of length 2 indicating the shape of the grid to
use for splitting the dataset. Default is c(2,2)}

\item{show_progress}{A boolean indicating if a plot should be displayed to
track the progressing of the calculation.}

\item{mindist}{The minimum length of a lixel. Defaut is NULL. If NULL,
mindist is set as lixel_length/10}
}
\value{
Vector of kernel densities (one for each origin)
}
\description{
The dataset is spatialy splitted in squares convering
the extent of the SpatiaLine. This greatly improve performance and memory
usage. To avoid frontier effect, a buffer is used (twice the size of the
range parameter) around the square when building local graphs. As a
consequence, too small squares will lead to longer calculation.
This version of the function can use multicore if a plan is defined
with the future package. If you do not use multicore, the regular function
nkde_grided will be slightely more efficient.
}
\examples{
data(mtl_network)
data(bike_accidents)
future::plan(future::multiprocess(workers=2))
lixels_nkde <- nkde_grided.mc(mtl_network, bike_accident,
      snap_dist = 150,
      lx_length = 150,
      mindist = 50,
      kernel_range = 800,
      kernel="quartic",
      grid_shape=c(5,5)
)
}
