% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde_execution_functions.R
\name{nkde_worker}
\alias{nkde_worker}
\title{The worker function for nkde and nkde.mc}
\usage{
nkde_worker(
  lines,
  events,
  samples,
  kernel_func,
  bw,
  method,
  div,
  digits,
  tol,
  max_depth,
  verbose = FALSE
)
}
\arguments{
\item{lines}{A SpatialLinesDataFrame with the sampling points. The
geoemtries must be a SpatialLinesDataFrame (may crash if some geometries
 are invalid)}

\item{events}{A SpatialPointsDataFrame representing the events on the
network. The points will be snapped on the network.}

\item{samples}{A SpatialPointsDataFrame representing the locations for
which the densities will be estimated.}

\item{kernel_func}{A kernel function, got from select_kernel}

\item{bw}{The kernel bandwidth (in meters)}

\item{method}{The method to use when calculating the NKDE, must be one of
simple / discontinuous / continuous (see details for more information)}

\item{div}{The divisor to use for the kernel. Must be "n" (the number of
events within the radius around each sampling point), "bw" (the bandwith)
"none" (the simple sum).}

\item{digits}{The number of digits to keep in the spatial coordinates. It
ensures that topology is good when building the network. Default is 3}

\item{tol}{When adding the events and the sampling points to the network,
the minimum distance between these points and the lines extremities. When
points are closer, they are added at the extermity of the lines.}

\item{max_depth}{when using the continuous method, a recursive function
is applied. When links in network are very short, this might lead to very
deep recursion and crash. To avoid it, it is possible to set here a maximym
recusion depth. Okabe recommand a value of 3.
approximation.}

\item{verbose}{A boolean, indicating if the function should print messages
about process.}
}
\value{
A numerci vector with the nkde values
}
\description{
The worker function for nkde and nkde.mc
}
\examples{
#This is an internal function, no example provided
}
