% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pair_dists}
\alias{pair_dists}
\title{pairwise distance between two vectors}
\usage{
pair_dists(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a matrix width dimenion l(x) * l(y)
}
\description{
pairwise distance between two vectors
}
\keyword{internal}
