% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{plot_graph}
\alias{plot_graph}
\title{function to plot a graph (usefull to check connectivity)}
\usage{
plot_graph(graph)
}
\arguments{
\item{graph}{A graph object (produced with build_graph)}
}
\description{
function to plot a graph (usefull to check connectivity)
}
\examples{
#This is an internal function, no example provided
}
