% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouring_sf.R
\name{prepare_elements_netlistw}
\alias{prepare_elements_netlistw}
\title{Data preparation for network_listw}
\usage{
prepare_elements_netlistw(is, grid, snapped_points, lines, maxdistance)
}
\arguments{
\item{is}{The indices of the quadras to use in the grid}

\item{grid}{A feature collection of polygons representing the quadras to split
calculus}

\item{snapped_points}{The start and end points snapped to the lines}

\item{lines}{The lines representing the network}

\item{maxdistance}{The maximum distance between two observation to considere
them as neighbours.}
}
\value{
A list of two elements : selected points and selected lines
}
\description{
Function to prepare selected points and selected lines during
  the process.
}
\examples{
#no example provided, this is an internal function
}
\keyword{internal}
