% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_functions.R
\name{randomize_distmatrix}
\alias{randomize_distmatrix}
\title{Points on network randomization}
\usage{
randomize_distmatrix(graph, edge_df, n, start_vert = NULL)
}
\arguments{
\item{graph}{An graph object from igraph}

\item{edge_df}{A DataFrame describing the edges}

\item{n}{The number of point}

\item{start_vert}{The vertices from wich the distances will be calculated.
if null, then the distances are calculated from the generated locations.}
}
\value{
A numeric matrix with the distances between points
}
\description{
Randomize location of points on a network.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
