% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_functions_sf.R
\name{randomize_distmatrix2}
\alias{randomize_distmatrix2}
\title{Points on network randomization simplified}
\usage{
randomize_distmatrix2(graph, edge_df, n, resolution, nsim, start_vert = NULL)
}
\arguments{
\item{graph}{An graph object from igraph}

\item{edge_df}{A DataFrame describing the edges}

\item{n}{The number of point}

\item{resolution}{The maximum size of the network edges}

\item{nsim}{The number of distance matrices to generate}

\item{start_vert}{The vertices from which the distances will be calculated.
if null, then the distances are calculated from the generated locations.}
}
\value{
A numeric matrix with the distances between points
}
\description{
Randomize location of points on a network.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
