% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{remove_loop_lines}
\alias{remove_loop_lines}
\title{Remove loops}
\usage{
remove_loop_lines(lines, digits)
}
\arguments{
\item{lines}{A sf object with linestring type geometries}

\item{digits}{An integer indicating the number of digits to keep for the
spatial coordinates}
}
\value{
A sf object with linestring type geometries
}
\description{
Remove from a sf object with linestring type geometries the lines that have the
same starting and ending point.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
