% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEW_k_functions_nettime_sf.R
\name{rev_matrix}
\alias{rev_matrix}
\title{Rervese the elements in a matrix}
\usage{
rev_matrix(mat)
}
\arguments{
\item{mat}{The matrix to reverse}
}
\value{
A matrix
}
\description{
reverse the order of the elements in a matrix both
column and row wise
}
\keyword{internal}
