% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{simple_lines}
\alias{simple_lines}
\title{LineString to simple Line}
\usage{
simple_lines(lines)
}
\arguments{
\item{lines}{The featue collection of linestrings to modify}
}
\value{
An featue collection of linestrings
}
\description{
Split the polylines of a feature collection of linestrings in simple
segments
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
