% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_nkde.R
\name{simple_tnkde}
\alias{simple_tnkde}
\title{Simple TNKDE algorithm}
\usage{
simple_tnkde(
  graph,
  events,
  samples,
  samples_time,
  obw_net,
  bws_net,
  obw_time,
  bws_time,
  kernel_func,
  nodes,
  edges,
  div
)
}
\arguments{
\item{graph}{a graph object from igraph representing the network}

\item{events}{a SpatialPointsDataFrame representing the events. It must be
snapped on the network, and be nodes of the network. A column vertex_id
must indicate for each event its corresponding node}

\item{samples}{a SpatialPointsDataFrame representing the sampling points.
The samples must be snapped on the network. A column edge_id must indicate
for each sample on which edge it is snapped.}

\item{samples_time}{a numeric vector indicating when the densities must be
sampled}

\item{obw_net}{a float, the overall network bandwidth, used to standardize densities if
div = "bw"}

\item{bws_net}{a vector indicating the network kernel bandwidth (in meters) for each
event}

\item{obw_time}{a float, the overall time bandwidth, used to standardize densities if
div = "bw"}

\item{bws_time}{a vector indicating the time kernel bandwidth for each
event}

\item{kernel_func}{a function obtained with the function select_kernel}

\item{nodes}{a SpatialPointsDataFrame representing the nodes of the network}

\item{edges}{a SpatialLinesDataFrame representing the edges of the network}

\item{div}{The divisor to use for the kernel. Must be "n" (the number of
events within the radius around each sampling point), "bw" (the bandwidth)
"none" (the simple sum).}
}
\value{
a list of two matrices. The first one ins the matrix of the densities,
the rows are the samples and the columns the time. The second has the same
dimensions and contains the number of events influencing each sample
}
\description{
Function to perform the simple tnkde.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
