% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{small_mtl_network}
\alias{small_mtl_network}
\title{Smaller subset road network of Montreal}
\format{
A sf object with 1244 rows and 2 variables
\describe{
  \item{TYPE}{the type of road}
  \item{geom}{the geometry (linestrings)}
}
}
\source{
<https://donnees.montreal.ca/ville-de-montreal/geobase>
}
\usage{
small_mtl_network
}
\description{
A feature collection (sf object) representing the road network of Montreal. The EPSG is 3797, and the data comes from the
Montreal OpenData website. It is only a small extract in central districts used to demonstrate the
main functions of spNetwork. It is mainly used internally for tests.
}
\keyword{datasets}
