% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{sp_char_index}
\alias{sp_char_index}
\title{Coordinates to unique character verctor}
\usage{
sp_char_index(coords, digits)
}
\arguments{
\item{coords}{A n * 2 matrix representing the coordinates}

\item{digits}{The number of digits to keep from the coordinates}
}
\value{
A vector character vector of length n
}
\description{
Generate a character vector based on a coodinates matrix and
the maximum number of digits to keep.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
