% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cpp_classes.R
\name{spatial_index$new}
\alias{spatial_index$new}
\title{Constructor method for a spatial_index object}
\arguments{
\item{x}{\itemize{
\item A numeric matrix with 4 columns (minX, minY, maxX, maxY)
}}
}
\value{
a new instance of the spatial_index class (c++ pointer)
}
\description{
An object (c++ pointer) that uses the geometry
index library from boost to perform spatial queries
}
