% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cpp_classes.R
\name{spatial_index$tree_request}
\alias{spatial_index$tree_request}
\title{spatial request on rtree index (internal)}
\arguments{
\item{reqBbox}{- A numeric vector with 4 values (minX, minY, maxX, maxY)}
}
\value{
an IntegerVector with the indices of the intersected boxes
}
\description{
A method to get the boxes in the rtree intersecting another box
}
