% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-SpatialNetwork.R
\docType{class}
\name{SpatialNetwork-class}
\alias{$,SpatialNetwork-method}
\alias{SpatialNetwork}
\alias{SpatialNetwork-class}
\alias{[[,SpatialNetwork,ANY,missing-method}
\alias{[[<-,SpatialNetwork,ANY,missing-method}
\title{The SpatialNetwork class and constructor function}
\usage{
SpatialNetwork(sl, g, weights, weightfield)
}
\arguments{
\item{sl}{object of one of (a sublasses of) \link{SpatialLines}}

\item{g}{object of class \link{igraph}}

\item{weights}{weight for edges (defaults to length of linestring)}

\item{weightfield}{character; name of the attribute field of \code{sl} that will be used as weights}
}
\value{
object of class \link{SpatialNetwork-class}
}
\description{
SpatialNetwork: a Class for Spatial Networks (lines and edges)
}
\details{
A class to store spatial networks, such as a road network.
}
\note{
note
}
\section{Slots}{
 
 \describe{
   \item{\code{bbox}}{matrix, holding bounding box}
   \item{\code{proj4string}}{object of class \link[sp]{CRS-class}}
   \item{\code{lines}}{list}
   \item{\code{data}}{data.frame, holding attributes associated with lines}
   \item{\code{g}}{object of a subclass of \link[igraph]{igraph}}
   \item{\code{weightfield}}{character; describing the weight field used}
 }
}
\examples{
library(sp) # Lines, SpatialLines
l0 = cbind(c(1, 2), c(0, 0))
l1 = cbind(c(0, 0, 0), c(0, 1, 2))
l2 = cbind(c(0, 0, 0), c(2, 3, 4))
l3 = cbind(c(0, 1, 2), c(2, 2, 3))
l4 = cbind(c(0, 1, 2), c(4, 4, 3))
l5 = cbind(c(2, 2), c(0, 3))
l6 = cbind(c(2, 3), c(3, 4))
LL = function(l, ID) Lines(list(Line(l)), ID)
l = list(LL(l0, "e"), LL(l1, "a"), LL(l2, "b"), LL(l3, "c"), LL(l4, "d"), LL(l5, "f"), LL(l6, "g"))
sl = SpatialLines(l)
sln = SpatialNetwork(sl)
plot(sln@g$x, sln@g$y, col = sln@g$n, pch = 16, cex = 2, asp = 1)
lines(sl)
library(igraph) # E
text(sln@g$x, sln@g$y, E(sln@g), pos = 4)
plot(sln@g)
}
\author{
Edzer Pebesma
}

