% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,SpatialNetwork-method}
\alias{points.SpatialNetwork}
\title{extract method for testClass}
\usage{
\S4method{[}{SpatialNetwork}(x, i, j, ..., drop = TRUE)

\method{points}{SpatialNetwork}(x, ..., col = "red", cex = 2)
}
\arguments{
\item{x}{object of class SpatialNetwork}

\item{i}{numeric; features to select}

\item{j}{numeric or character; attributes to select}

\item{...}{ignored}

\item{drop}{logical; ignored}

\item{col}{color}

\item{cex}{symbol size}
}
\description{
extract method for testClass
}

