% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde_execution_functions.R
\name{split_by_grid}
\alias{split_by_grid}
\title{Function to split the dataset according to a grid}
\usage{
split_by_grid(grid, samples, events, lines, bw, tol, digits)
}
\arguments{
\item{grid}{a spatial grid to split the data within}

\item{samples}{A spatialPointsDataFrame of the samples points}

\item{events}{A spatialPointsDataFrame of the events points}

\item{lines}{A SpatialLinesDataFrame representing the network}

\item{bw}{the kernel bandwidth (used to avoid egde effect)}

\item{tol}{a float indicating the spatial tolerance when snapping events on
lines}

\item{digits}{the number of digits to keep}
}
\value{
a list with the splitted dataset
}
\description{
Function to split the dataset according to a grid
}
\examples{
#This is an internal function, no example provided
}
