% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cpp_classes.R
\docType{package}
\name{spNetwork-package}
\alias{spNetwork}
\alias{spNetwork-package}
\title{spNetwork: Spatial Analysis on Network}
\description{
Perform spatial analysis on network. Implement several methods for spatial analysis on network: Network Kernel Density estimation, building of spatial matrices based on network distance ('listw' objects from 'spdep' package), K functions estimation for point pattern analysis on network, k nearest neighbours on network, reachable area calculation, and graph generation References: Okabe et al (2019) \doi{10.1080/13658810802475491}; Okabe et al (2012, ISBN:978-0470770818);Baddeley et al (2015, ISBN:9781482210200).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jeremygelb.github.io/spNetwork/}
  \item Report bugs at \url{https://github.com/JeremyGelb/spNetwork/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Gelb \email{jeremy.gelb@ucs.inrs.ca} (\href{https://orcid.org/0000-0002-7114-2714}{ORCID})

Other contributors:
\itemize{
  \item Philippe Apparicio \email{philippe.apparicio@ucs.inrs.ca} (\href{https://orcid.org/0000-0001-6466-9342}{ORCID}) [contributor]
}

}
