% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indexing_sf.R
\name{st_bbox_by_feature}
\alias{st_bbox_by_feature}
\title{Obtain all the bounding boxes of a feature collection}
\usage{
st_bbox_by_feature(x)
}
\arguments{
\item{x}{a feature collection}
}
\value{
a matrix (xmin, ymin, xmax, ymax)
}
\description{
Obtain all the bounding boxes of a feature collection (INTERNAL).
}
\examples{
#This is an internal function, no example provided
}
