% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{st_bbox_geom}
\alias{st_bbox_geom}
\title{sf geometry bbox}
\usage{
st_bbox_geom(x)
}
\arguments{
\item{x}{A feature collection}
}
\value{
A feature collection of polygons
}
\description{
Generate polygon as the bounding box of a feature collection
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
