% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{surrounding_points}
\alias{surrounding_points}
\title{Points along polygon boundary}
\usage{
surrounding_points(polygons, dist)
}
\arguments{
\item{polygons}{A feature collection of polygons}

\item{dist}{The distance between the points}
}
\value{
A feature collection of points representing the points arrond the polygond
}
\description{
Generate a feature collection of points by placing points along the border of
polygons of a feature collection.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
