% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{theatres_mtl}
\alias{theatres_mtl}
\title{Theatres}
\format{
A SpatialPointsDataFrame with 54 rows and 2 variables
}
\source{
\url{http://donnees.ville.montreal.qc.ca/}
}
\usage{
theatres_mtl
}
\description{
A SpatialPointsDataFrame representing some theatres in Montreal
}
\keyword{datasets}
