% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{triangle_kernel}
\alias{triangle_kernel}
\title{triangle kernel}
\usage{
triangle_kernel(d, bw)
}
\arguments{
\item{d}{the distance from the event}

\item{bw}{the bandwidth used for the kernel}
}
\value{
the estimated density
}
\description{
triangle kernel
}
\examples{
#This is an internal function, no example provided
}
