% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{tricube_kernel}
\alias{tricube_kernel}
\title{tricube kernel}
\usage{
tricube_kernel(d, bw)
}
\arguments{
\item{d}{the distance from the event}

\item{bw}{the bandwidth used for the kernel}
}
\value{
the estimated density
}
\description{
tricube kernel
}
\examples{
#This is an internal function, no example provided
}
