% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tricube_kernelos}
\alias{tricube_kernelos}
\title{c++ tricube kernel for one distance}
\usage{
tricube_kernelos(d, bw)
}
\arguments{
\item{d}{a double, the distances for which the density must be calculated}

\item{bw}{a double representing the size of the kernel bandwidth}
}
\description{
c++ tricube kernel for one distance
}
\keyword{internal}
