% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrones_sf.R
\name{trim_lines_at}
\alias{trim_lines_at}
\title{Helper for isochrones lines cutting}
\usage{
trim_lines_at(df1, graph_result, d, dd, i, donught)
}
\arguments{
\item{df1}{A features collection of linestrings with some specific fields.}

\item{graph_result}{A list produced by the functions build_graph_directed or build_graph.}

\item{d}{the end distance of this isochrones.}

\item{dd}{the start distance of this isochrones.}

\item{i}{the actual iteration.}

\item{donught}{A boolean indicating if the returned isochrone will be plained or a donught.}
}
\value{
A feature collection of lines
}
\description{
last operation for isochrone calculation, cutting the lines at
their begining and ending. This is a worker function for calc_isochrones.
}
\keyword{internal}
