% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{triweight_kernel_cpp}
\alias{triweight_kernel_cpp}
\title{c++ triweight kernel}
\usage{
triweight_kernel_cpp(d, bw)
}
\arguments{
\item{d}{a vector of distances for which the density must be calculated}

\item{bw}{a double representing the size of the kernel bandwidth}
}
\description{
c++ triweight kernel
}
\keyword{internal}
