% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{weightfield}
\alias{weightfield}
\alias{weights}
\alias{weights,SpatialNetwork-method}
\alias{weights<-}
\alias{weights<-,SpatialNetwork,ANY,character-method}
\alias{weights<-,SpatialNetwork,ANY,vector-method}
\alias{weights<-,SpatialNetwork,character,vector-method}
\alias{weights<-,SpatialNetwork,character-method,}
\alias{weights<-,SpatialNetwork,vector-method,}
\title{Get or set weight field in SpatialNetwork}
\usage{
weights(x)

\S4method{weights}{SpatialNetwork}(x)

\S4method{weights}{SpatialNetwork,ANY,character}(x) <- value

\S4method{weights}{SpatialNetwork,ANY,vector}(x) <- value

\S4method{weights}{SpatialNetwork,character,vector}(x, weightfield) <- value
}
\arguments{
\item{x}{SpatialNetwork to use}

\item{value}{Either the name of the variable to use as the weight field or
a vector containing the weights to use if \code{varname} is
passed to the replacement function.}

\item{weightfield}{The name of the variable to set/use.}
}
\description{
Get or set value of weight field in SpatialNetwork
}
\section{Details}{

These functions manipulate the value of weightfield in a
SpatialNetwork. When changing the value of weightfield, the weights
of the graph network are updated with the values of the corresponding
variables.
}

