% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{add_center_lines}
\alias{add_center_lines}
\title{Add to each line of a SpatialLinesDataFrame an additionnal vertex at its
center}
\usage{
add_center_lines(lines, show_progress = T)
}
\arguments{
\item{lines}{The SpatialLinesDataframe to use}

\item{show_progress}{A boolean indicating if a progress bar must be displayed}
}
\value{
An object of class SpatialLinesDataframe (package sp)
}
\description{
Add to each line of a SpatialLinesDataFrame an additionnal vertex at its
center
}
\examples{
#This is an internal function, no example provided
}
