% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{add_center_lines}
\alias{add_center_lines}
\title{Add center vertex to lines}
\usage{
add_center_lines(lines)
}
\arguments{
\item{lines}{The SpatialLinesDataframe to use}
}
\value{
An object of class SpatialLinesDataframe (package sp)
}
\description{
Add to each line of a SpatialLinesDataFrame an additional vertex at its
center.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
