% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{add_vertices_lines}
\alias{add_vertices_lines}
\title{Add vertices to a feature collection of linestrings}
\usage{
add_vertices_lines(lines, points, nearest_lines_idx, mindist)
}
\arguments{
\item{lines}{The feature collection of linestrings to modify}

\item{points}{The feature collection of points to add to as vertex to the lines}

\item{nearest_lines_idx}{For each point, the index of the nearest line}

\item{mindist}{The minimum distance between one point and the extremity of
the line to add the point as a vertex.}
}
\value{
A feature collection of linestrings
}
\description{
Add vertices (feature collection of points) to their nearest lines
  (feature collection of linestrings), may fail if the line geometries are self intersecting.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
