% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{add_vertices_lines}
\alias{add_vertices_lines}
\title{Add vertices (SpatialPoints) to many lines (SpatialLines), may fail
if the lines geometries are self intersecting}
\usage{
add_vertices_lines(lines, points, tol = 0.1, check = TRUE, show_progress = T)
}
\arguments{
\item{lines}{The SpatialLinesDataframe to modify}

\item{points}{The SpatialPoints to add to as vertex to the lines}

\item{tol}{The max distance to between lines and points so that they
are added}
}
\value{
An object of class SpatialLinesDataFrame (package sp)
}
\description{
Add vertices (SpatialPoints) to many lines (SpatialLines), may fail
if the lines geometries are self intersecting
}
\examples{
#This is an internal function, no example provided
}
