% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde_execution_functions_sf.R
\name{aggregate_points}
\alias{aggregate_points}
\title{Events aggregation}
\usage{
aggregate_points(points, maxdist, weight = "weight")
}
\arguments{
\item{points}{The feature collection of points to contract (must have a weight column)}

\item{maxdist}{The distance to use}

\item{weight}{The name of the column to use as weight (default is "weight").
The values of the aggregated points for this column will be summed. For all
the other columns, only the max value is retained.}
}
\value{
A new feature collection of points
}
\description{
Function to aggregate points within a radius.
}
\details{
This function can be used to aggregate points within a radius. This
  is done by using the dbscan algorithm. This process is
  repeated until no more modification is applied.
}
\examples{
data(bike_accidents)
bike_accidents$weight <- 1
agg_points <- aggregate_points(bike_accidents, 5)
}
