% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{bike_accidents}
\alias{bike_accidents}
\title{An extract of the bike accidents on Montreal road network}
\format{
A SpatialPointsDataFrame with 2946 rows and 4 variables: \describe{
  \item{LOC_X}{The X coordinates of the accident (EPSG : 32188)}
  \item{LOC_Y}{The Y coordinates of the accident (EPSG : 32188)}
  \item{NB_VICTIME}{the number of victims} \item{AN}{the year of the
  accident} ... }
}
\source{
\url{http://donnees.ville.montreal.qc.ca/}
}
\usage{
bike_accidents
}
\description{
A SpatialPointsDataFrame representing bike accidents that occured on the
montreal road network in 2016
}
\keyword{datasets}
