% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{build_grid}
\alias{build_grid}
\title{Spatial grid}
\usage{
build_grid(grid_shape, spatial)
}
\arguments{
\item{grid_shape}{A numeric vector of length 2 indicating the number
of rows and the numbers of columns of the grid}

\item{spatial}{A list of spatial feature collections objects (package sf)}
}
\value{
A feature collection of polygons representing the grid
}
\description{
Generate a grid of a specified shape in the bbox of a Spatial object.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
