% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indexing.R
\name{build_quadtree}
\alias{build_quadtree}
\title{Build a quadtree}
\usage{
build_quadtree(data)
}
\arguments{
\item{data}{a SpatialLinesDataFrame or a SpatialPointsDataFrame}
}
\value{
quad tree object from package SearchTrees
}
\description{
Generate quad tree object from package SearchTrees, useful to speed up
spatial requesting.
}
\examples{
#This is an internal function, no example provided
}
