% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indexing_sf.R
\name{build_quadtree}
\alias{build_quadtree}
\title{Build a quadtree}
\usage{
build_quadtree(data)
}
\arguments{
\item{data}{a feature collection of linestrings or a feature collection of points}
}
\value{
quadtree object from package SearchTrees
}
\description{
Generate a quadtree object from package SearchTrees, useful to speed up
spatial requesting (INTERNAL).
}
\examples{
#This is an internal function, no example provided
}
