% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde_execution_functions.R
\name{check_geometries}
\alias{check_geometries}
\title{Function to check if the geometries given by the user are valide}
\usage{
check_geometries(lines, samples, events)
}
\arguments{
\item{lines}{a SpatialLinesDataFrame}

\item{samples}{a SpatialPointsDataFrame of the samples}

\item{events}{a SpatialPointsDataFrame of the events}
}
\value{
TRUE if all the checks are passed
}
\description{
Function to check if the geometries given by the user are valide
}
\examples{
#This is an internal function, no example provided
}
