% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde_execution_functions.R
\name{clean_events}
\alias{clean_events}
\title{Clean events geometries}
\usage{
clean_events(events, digits = 5, agg = NULL)
}
\arguments{
\item{events}{The SpatialPointsDataFrame to contract (must have a weight column)}

\item{digits}{The number of digits to keep}

\item{agg}{A double indicating if the points must be aggregated within a distance.
if NULL, then the points are aggregated by rounding the coordinates.}
}
\value{
A new SpatialPointsDataFrame
}
\description{
Function to avoid having events at the same location.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
