% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indexing.R
\name{closest_points}
\alias{closest_points}
\title{build a quad tree index and solve the nearest neighbour problem for two
SpatialPointsDataFrame}
\usage{
closest_points(origins, targets)
}
\arguments{
\item{origins}{a SpatialPointsDataFrame}

\item{targets}{a SpatialPointsDataFrame}
}
\value{
for each origin point, the index of the nearest target point
}
\description{
build a quad tree index and solve the nearest neighbour problem for two
SpatialPointsDataFrame
}
\examples{
#This is an internal function, no example provided
}
