% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indexing_sf.R
\name{closest_points}
\alias{closest_points}
\title{Find closest points}
\usage{
closest_points(origins, targets)
}
\arguments{
\item{origins}{a feature collection of points}

\item{targets}{a feature collection of points}
}
\value{
for each origin point, the index of the nearest target point
}
\description{
Solve the nearest neighbour problem for two feature collections of points
This is a simple wrap-up of the dbscan::kNN function
}
\examples{
data(mtl_libraries)
data(mtl_theatres)
close_libs <- closest_points(mtl_theatres, mtl_libraries)
}
