% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{ess_kernel}
\alias{ess_kernel}
\title{Worker for simple NKDE algorithm}
\usage{
ess_kernel(
  graph,
  y,
  bw,
  kernel_func,
  samples,
  nodes,
  edges,
  sample_tree,
  edges_tree
)
}
\arguments{
\item{graph}{a graph object from igraph representing the network}

\item{y}{the index of the actual event}

\item{bw}{a float indicating the kernel bandwidth (in meters)}

\item{kernel_func}{a function obtained with the function select_kernel}

\item{samples}{a SpatialPointsDataFrame representing the sampling points. The
samples must be snapped on the network. A column edge_id must indicate for
each sample on which edge it is snapped.}

\item{nodes}{a SpatialPointsDataFrame representing the nodes of the network}

\item{edges}{a SpatialLinesDataFrame representing the edges of the network}

\item{sample_tree}{a quadtree object, the spatial index of the samples}

\item{edges_tree}{a quadtree object, the spatial index of the edges}
}
\description{
The worker function to perform the simple nkde.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
