% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_nkde.R
\name{exe_nkde}
\alias{exe_nkde}
\title{Worker function for the nkde_grided.mc function.}
\usage{
exe_nkde(
  is,
  grid,
  lines,
  lixels,
  points,
  kernel,
  kernel_range,
  snap_dist,
  digits,
  tol,
  direction,
  lixels_tree,
  lines_tree,
  points_tree
)
}
\arguments{
\item{is}{The row indices of the quadra to use in the grid}

\item{grid}{The SpatialPolygons object reprsenting the grid for splitted
calculation}

\item{lines}{The SpatialLinesDataFrame to use as a network}

\item{lixels}{The SpatialLinesDataFrame representing the lixels}

\item{points}{The events to use in the kernel density estimation}

\item{kernel}{The name of the kernel function to use (must be one of
quartic, gaussian or epanechnikov). Default is Quartic}

\item{kernel_range}{The range of the kernel function}

\item{snap_dist}{The maximum distance snapping between points and lines}

\item{digits}{The number of digits to keep in the coordinates of the
geometries}

\item{tol}{The tolerence for topological operations}

\item{direction}{indicate a field giving informations about authorized
traveling direction on lines. if NULL, then all lines can be used in both
directions. Must be the name of a column otherwise. The values of the
column must be "FT" (From - To), "TF" (To - From) or "Both".}

\item{lixels_tree}{a quad tree index for the lixels}

\item{lines_tree}{a quad tree index for the lines}

\item{points_tree}{a quad tree index for the points}
}
\value{
A SpatialLinesDataFrame of the lines and their kernel densities
}
\description{
Worker function for the nkde_grided.mc function.
}
\examples{
#This is an internal function, no example provided
}
