% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{find_vertices}
\alias{find_vertices}
\title{function to match some points (SpatialPointsDataFrame) to the vertices of
a graph}
\usage{
find_vertices(spvertices, points, digits, tol = 0.1)
}
\arguments{
\item{spvertices}{The spatial vertices of a graph (produced whith
build_graph)}

\item{points}{the SpatialPointsDataFrame to match}

\item{digits}{The number of digits to keep from the coordinates}

\item{tol}{the maximum distance between a point and a vertex}
}
\value{
A vector of the corresponding vertices id, multiple points may
belong to the same vertex
}
\description{
function to match some points (SpatialPointsDataFrame) to the vertices of
a graph
}
\examples{
#This is an internal function, no example provided
}
