% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numeric values}

\item{na.rm}{A boolean indicating if we filter the NA values}
}
\value{
The geometric mean of x
}
\description{
Function to calculate the geometric mean.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
