% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_functions.R
\name{kfunctions}
\alias{kfunctions}
\title{Network k and g functions}
\usage{
kfunctions(
  lines,
  points,
  start,
  end,
  step,
  width,
  nsim,
  conf_int = 0.05,
  digits = 2,
  tol = 0.1,
  agg = NULL,
  verbose = TRUE
)
}
\arguments{
\item{lines}{A SpatialLinesDataFrame with the sampling points. The geometries
must be a SpatialLinesDataFrame (may crash if some geometries are invalid)}

\item{points}{A SpatialPointsDataFrame representing the points on the
network. These points will be snapped on the network.}

\item{start}{A double, the start value for evaluating the k and g functions}

\item{end}{A double, the last value for evaluating the k and g functions}

\item{step}{A double, the jump between two evaluations of the k and g
function}

\item{width}{The width of each donut for the g-function}

\item{nsim}{An integer indicating the number of Monte Carlo simulations
required}

\item{conf_int}{A double indicating the width confidence interval (default =
0.05)}

\item{digits}{An integer indicating the number of digits to retain for the
spatial coordinates}

\item{tol}{When adding the points to the network, specify the minimum
distance between these points and the lines' extremities. When points are
closer, they are added at the extremity of the lines.}

\item{agg}{A double indicating if the events must be aggregated within a
distance. If NULL, the events are aggregated by rounding the coordinates.}

\item{verbose}{A Boolean indicating if progress messages should be displayed}
}
\value{
A list with the following values : \cr \itemize{ \item{plotk}{A
  ggplot2 object representing the values of the k-function} \item{plotg}{A
  ggplot2 object representing the values of the g-function} \item{values}{A
  DataFrame with the values used to build the plots} }
}
\description{
Calculate the k and g functions for a set of points on a
  network.
}
\details{
The k-function is a method to characterize the dispersion of a set
  of points. For each point, the numbers of other points in subsequent radii
  are calculated. This empirical k-function can be more or less clustered
  than a k-function obtained if the points were randomly located in space. In
  a network, the network distance is used instead of the Euclidean distance.
  This function uses Monte Carlo simulations to assess if the points are
  clustered or dispersed, and gives the results as a line plot. If the line
  of the observed k-function is higher than the shaded area representing the
  values of the simulations, then the points are more clustered than what we
  can expect from randomness and vice-versa. The function also calculates the
  g-function, a modified version of the k-function using rings instead of
  disks. The width of the ring must be chosen. The main interest is to avoid
  the cumulative effect of the classical k-function.
}
\examples{
data("libraries_mtl")
data("main_network_mtl")
result <- kfunctions(main_network_mtl, libraries_mtl,
     start = 0, end = 2500, step = 10,
     width = 200, nsim = 50,
     conf_int = 0.05, tol = 0.1, agg = NULL,
     verbose = FALSE)
}
