% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{lines_center}
\alias{lines_center}
\title{Centre points of lines}
\usage{
lines_center(lines)
}
\arguments{
\item{lines}{A feature collection of linestrings to use}
}
\value{
A feature collection of points
}
\description{
Generate a feature collection of points at the centre of the
lines of a feature collection of linestrings. The length of the lines is
used to determine their centres.
}
\examples{
\donttest{
data(mtl_network)
centers <- lines_center(mtl_network)
}
}
