% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{lines_center}
\alias{lines_center}
\title{Center points of lines}
\usage{
lines_center(lines)
}
\arguments{
\item{lines}{The SpatialLinesDataframe to use}
}
\value{
An object of class SpatialPointsDataFrame (package sp)
}
\description{
Generate a SpatialPointsDataFrame with line center points. The
  points are located at center of the line based on the length of the line.
}
\examples{
\dontrun{
networkgpkg <- system.file("extdata", "networks.gpkg", package = "spNetwork", mustWork = TRUE)
mtl_network <- rgdal::readOGR(networkgpkg,layer="mtl_network", verbose=FALSE)
centers <- lines_center(mtl_network)
}
}
