% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{lines_center}
\alias{lines_center}
\title{Generate a SpatialPointsDataFrame with line center points. The points are
located at middle of the line based on the length of the line}
\usage{
lines_center(lines)
}
\arguments{
\item{lines}{The SpatialLinesDataframe to use}
}
\value{
An object of class SpatialPointsDataFrame (package sp)
}
\description{
Generate a SpatialPointsDataFrame with line center points. The points are
located at middle of the line based on the length of the line
}
\examples{
data('mtl_network')
centers <- lines_center(mtl_network)
}
