% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{lines_direction}
\alias{lines_direction}
\title{Unify lines direction}
\usage{
lines_direction(lines, field)
}
\arguments{
\item{lines}{A SpatialLinesDataFrame}

\item{field}{Indicate a field giving informations about authorized
traveling direction on lines. if NULL, then all lines can be used in both
directions. Must be the name of a column otherwise. The values of the
column must be "FT" (From - To), "TF" (To - From) or "Both".}
}
\value{
A SpatialLinesDataFrame
}
\description{
A function to deal with the directions of lines.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
