% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{lines_extremities}
\alias{lines_extremities}
\title{Get lines extremities}
\usage{
lines_extremities(lines)
}
\arguments{
\item{lines}{A SpatialLinesDataFrame}
}
\value{
A SpatialPointsDataFrame
}
\description{
Generate a SpatialPointsDataFrame with the first and last vertex of each
line in a SpatialLinesDataFrame.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
