% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{lines_extremities}
\alias{lines_extremities}
\title{Get lines extremities}
\usage{
lines_extremities(lines)
}
\arguments{
\item{lines}{A feature collection of linestrings (simple Linestrings)}
}
\value{
A feature collection of points
}
\description{
Generate a feature collection of points with the first and last vertex
  of each line in a feature collection of linestrings.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
