% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{lines_points_along}
\alias{lines_points_along}
\title{Points along lines}
\usage{
lines_points_along(lines, dist)
}
\arguments{
\item{lines}{A feature collection of linestrings to use}

\item{dist}{The distance between the points along the lines}
}
\value{
A feature collection of points
}
\description{
Generate a feature collection of points along the lines of
feature collection of Linestrings.
}
\examples{
\donttest{
data(mtl_network)
new_pts <- lines_points_along(mtl_network,50)
}
}
