% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{lines_points_along}
\alias{lines_points_along}
\title{Points along lines}
\usage{
lines_points_along(lines, dist)
}
\arguments{
\item{lines}{The SpatialLinesDataframe to use}

\item{dist}{The distance between the points along the lines}
}
\value{
An object of class SpatialLinesDataframe (package sp)
}
\description{
Generate points along the lines of a SpatialLinesDataFrame.
}
\examples{
\dontrun{
#data(mtl_network)
networkgpkg <- system.file("extdata", "networks.gpkg", package = "spNetwork", mustWork = TRUE)
mtl_network <- rgdal::readOGR(networkgpkg,layer="mtl_network", verbose=FALSE)
new_pts <- lines_points_along(mtl_network,50)
}
}
