% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{main_network_mtl}
\alias{main_network_mtl}
\title{An extract of the Montreal road network with only the main roads}
\format{
A SpatialLinesDataFrame with 17304 rows and 3 variables: \describe{
  \item{CLASS}{The category of the road} \item{NAME}{The name of the road}
  \item{TYPE}{the type of the road} }
}
\source{
\url{http://donnees.ville.montreal.qc.ca/}
}
\usage{
main_network_mtl
}
\description{
A SpatialLinesDataFrame representing the main roads of the Montreal road
network.
}
\keyword{datasets}
