% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mtl_libraries}
\alias{mtl_libraries}
\title{Libraries of Montreal}
\format{
A sf object with 55 rows and 3 variables.
\describe{
  \item{CP}{the postal code}
  \item{NAME}{the name of the library}
  \item{geom}{the geometry (points)}
}
}
\source{
<https://donnees.montreal.ca/ville-de-montreal/lieux-culturels>
}
\usage{
mtl_libraries
}
\description{
A feature collection (sf object) representing the libraries of Montreal. The EPSG is 3797 and the data comes from the
Montreal OpenData website.
}
\keyword{datasets}
