% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mtl_network}
\alias{mtl_network}
\title{Road network of Montreal}
\format{
A sf object with 2945 rows and 2 variables
\describe{
  \item{ClsRte}{the category of the road}
  \item{geom}{the geometry (linestrings)}
}
}
\source{
<https://donnees.montreal.ca/ville-de-montreal/geobase>
}
\usage{
mtl_network
}
\description{
A feature collection (sf object) representing the road network of Montreal. The EPSG is 3797, and the data comes from the
Montreal OpenData website. It is only a small subset in central districts used to demonstrate the
main functions of spNetwork.
}
\keyword{datasets}
