% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde_execution_functions.R
\name{nkde.mc}
\alias{nkde.mc}
\title{Network Kernel density estimate}
\usage{
nkde.mc(
  lines,
  events,
  w,
  samples,
  kernel_name,
  bw,
  method,
  div = "bw",
  max_depth = 15,
  digits = 5,
  tol = 0.1,
  grid_shape = c(1, 1),
  verbose = TRUE
)
}
\arguments{
\item{lines}{A SpatialLinesDataFrame with the sampling points. The
geoemtries must be a SpatialLinesDataFrame (may crash if some geometries
 are invalid)}

\item{events}{A SpatialPointsDataFrame representing the events on the
network. The points will be snapped on the network.}

\item{w}{A vector representing the weight of each event}

\item{samples}{A SpatialPointsDataFrame representing the locations for
which the densities will be estimated}

\item{kernel_name}{The name of the kernel to use. Must be one of triangle,
gaussian, tricube, cosine ,triweight, quartic, or epanechnikov.}

\item{bw}{The kernel bandwidth (in meters)}

\item{method}{The method to use when calculating the NKDE, must be one of
simple / discontinuous / continuous (see details for more information)}

\item{div}{The divisor to use for the kernel. Must be "n" (the number of
events within the radius around each sampling point), "bw" (the bandwith)
"none" (the simple sum).}

\item{max_depth}{when using the continuous method, a recursive function
is applied. When links in network are very short, this might lead to very
deep recursion and crash. To avoid it, it is possible to set here a maximym
recusion depth. Okabe recommand a value of 3, but it highly depends on the
length of the lines of the network. 15 should be a good approximation in
most cases, but might be time consuming.}

\item{digits}{The number of digits to keep in the spatial coordinates. It
ensures that topology is good when building the network. Default is 3}

\item{tol}{When adding the events and the sampling points to the network,
the minimum distance between these points and the lines extremities. When
points are closer, they are added at the extermity of the lines.}

\item{grid_shape}{A vector of two values indicating how the study area
must be splitted when performing the calculus (see details). Defaut is c(1,1)}

\item{verbose}{A boolean, indicating if the function should print messages
about process.}
}
\value{
A vector of values, they are the density estimates at samplings
points
}
\description{
Calculate the Network Kernel Density Estimate based on a network of lines,
sampling points, and events. This version can use the current defined plan
with the package future. This slightly increase the calculus to do, but can
increase speed a lot by dividing the burden on multiple cores.
}
\details{
for details, please see the function nkde
}
