% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde.R
\name{nkde}
\alias{nkde}
\title{nkde function.}
\usage{
nkde(
  lines,
  points,
  snap_dist,
  lx_length,
  lixels = NULL,
  line_weight = "length",
  kernel_range,
  kernel = "quartic",
  tol = 0.1,
  digits = 3,
  mindist = NULL,
  weights = NULL,
  grid_shape = c(1, 1),
  verbose = "all"
)
}
\arguments{
\item{lines}{The SpatialLinesDataFrame to use as a network}

\item{points}{The events to use in the kernel density estimation}

\item{snap_dist}{The maximum distance snapping between points and lines}

\item{lx_length}{The expected length of a lixel}

\item{lixels}{A SpatialLinesDataFrame created before with the same network.
If NULL, then the lixels will be created using the argument lx_lenght and
mindist}

\item{line_weight}{The ponderation to use for lines. Default is "length"
(the geographical length), but can be the name of a column. The value is
considered proportional with the geographical length of the lines.}

\item{kernel_range}{The range of the kernel function}

\item{kernel}{The name of the kernel function to use (must be one of
quartic, gaussian or epanechnikov). Default is Quartic}

\item{tol}{The tolerence for topological operations}

\item{digits}{The number of digits to keep in the coordinates of the
geometries}

\item{mindist}{The minimum length of a lixel. Defaut is NULL. If NULL,
mindist is set as lx_length/10}

\item{weights}{The name of a column of the SpatialPointsDataFrame
containing the weight of each point. Default is NULL. If NULL then each
point has a weight of 1.}

\item{grid_shape}{A vector of length 2 indicating the shape of the grid to
use for splitting the dataset. Default is c(1,1)}

\item{verbose}{A string indicating how the advance of the process is
displayed. Default is all. "text" show only the main steps, "progressbar"
show main steps and progress bar for intermediar steps, "all" add a plot
of the grid to follow visually the processing, "silent" show nothing}
}
\value{
Vector of kernel densities (one for each origin)
}
\description{
The dataset is spatialy splitted in squares convering the extent of the
SpatiaLine. This greatly improve performance and memory usage. To avoid
frontier effect, a buffer is used (size of the kernel range parameter)
around the square when building local graphs. As a consequence, too small
squares will lead to longer calculation.
}
\examples{
data(mtl_network)
data(bike_accidents)
lixels_nkde <- nkde(mtl_network, bike_accidents,
      snap_dist = 150,
      lx_length = 150,
      mindist = 50,
      kernel_range = 800,
      kernel='quartic',
      grid_shape=c(5,5)
)
}
