% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nkde_get_loo_values}
\alias{nkde_get_loo_values}
\title{The exposed function to calculate NKDE likelihood cv}
\usage{
nkde_get_loo_values(
  method,
  neighbour_list,
  sel_events,
  sel_events_wid,
  events,
  events_wid,
  weights,
  bws_net,
  kernel_name,
  line_list,
  max_depth,
  min_tol,
  cvl
)
}
\arguments{
\item{method}{a string, one of "simple", "continuous", "discontinuous"}

\item{neighbour_list}{a List, giving for each node an IntegerVector with
its neighbours}

\item{events}{a NumericVector indicating the nodes in graph beeing events}

\item{weights}{a matrix with the weights associated with each events (row) for each
bws_net (cols).}

\item{bws_net}{an arma::vec with the network bandwidths to consider}

\item{kernel_name}{a string with the name of the kernel to use}

\item{line_list}{a DataFrame describing the lines}

\item{max_depth}{the maximum recursion depth}

\item{min_tol}{a double indicating by how much 0 in densities values must be replaced}

\item{cvl}{a boolean indicating if the Cronie (TRUE) or CV likelihood (FALSE) must be used}
}
\value{
a matrix with the CV score for each pair of bandiwdths
}
\description{
The exposed function to calculate NKDE likelihood cv
}
\examples{
# no example provided, this is an internal function
}
