% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions_sf.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot graph}
\usage{
plot_graph(graph)
}
\arguments{
\item{graph}{A graph object (produced with build_graph)}
}
\description{
Function to plot a graph (useful to check connectivity).
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
