% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{quartic_kernel}
\alias{quartic_kernel}
\title{quartic kernel}
\usage{
quartic_kernel(d, bw)
}
\arguments{
\item{d}{the distance from the event}

\item{bw}{the bandwidth used for the kernel}
}
\value{
the estimated density
}
\description{
quartic kernel
}
\examples{
#This is an internal function, no example provided
}
