% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions_sf.R
\name{quartic_kernel}
\alias{quartic_kernel}
\title{Quartic kernel}
\usage{
quartic_kernel(d, bw)
}
\arguments{
\item{d}{The distance from the event}

\item{bw}{The bandwidth used for the kernel}
}
\value{
The estimated density
}
\description{
Function implementing the quartic kernel.
}
\examples{
#This is an internal function, no example provided
}
