% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quartic_kernel_cpp}
\alias{quartic_kernel_cpp}
\title{c++ quartic kernel}
\usage{
quartic_kernel_cpp(d, bw)
}
\arguments{
\item{d}{a vector of distances for which the density must be calculated}

\item{bw}{a double representing the size of the kernel bandwidth}
}
\description{
c++ quartic kernel
}
\keyword{internal}
