% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{reverse_lines}
\alias{reverse_lines}
\title{Reverse lines}
\usage{
reverse_lines(lines)
}
\arguments{
\item{lines}{A SpatialLinesDataFrame}
}
\value{
A SpatialLinesDataFrame
}
\description{
A function to reverse the order of the vertices of lines
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
