% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouring.R
\name{select_dist_function}
\alias{select_dist_function}
\title{Select the distance to weight function}
\usage{
select_dist_function(dist_func = "inverse")
}
\arguments{
\item{dist_func}{Could be a name in c('inverse', 'identity',
'squared inverse') or a function with only one parameter x}
}
\value{
A vectorized function used to convert distance into spatial weights
}
\description{
Select a function to convert distance to weights
if a function is provided, this function will be vectorized.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
