% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{select_kernel}
\alias{select_kernel}
\title{select the right kernel function with its name}
\usage{
select_kernel(name)
}
\arguments{
\item{name}{The name of the kernel to use}
}
\value{
A kernel function
}
\description{
select the right kernel function with its name
}
\examples{
#This is an internal function, no example provided
}
