% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions_sf.R
\name{simple_lines}
\alias{simple_lines}
\title{LineString to simple Line}
\usage{
simple_lines(lines)
}
\arguments{
\item{lines}{The featue collection of linestrings to modify}
}
\value{
An featue collection of linestrings
}
\description{
Split the polylines of a feature collection of linestrings in simple
segments at each vertex. The values of the columns are duplicated for each segment.
}
\examples{
\donttest{
data(mtl_lines)
new_lines <- simple_lines(mtl_lines)
}
}
