% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{simple_lines}
\alias{simple_lines}
\title{LineString to simple Line}
\usage{
simple_lines(lines)
}
\arguments{
\item{lines}{The SpatialLinesDataframe to modify}
}
\value{
An object of class SpatialLinesDataFrame (package sp)
}
\description{
Split the polylines of a SpatialLinesDataFrame object in simple
  lines.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
