% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{small_mtl_network}
\alias{small_mtl_network}
\title{A smaller extract of the Montreal road network}
\format{
A SpatialLinesDataFrame with 1244 rows and 4 variables:
\describe{
  \item{IdRte}{a unique id for each road}
  \item{ClsRte}{the type of the road}
  \item{ClsFnction}{the function of the road}
  \item{nbAccident}{the number of
  bike accidents recorded on the road} ... }
}
\source{
\url{http://donnees.ville.montreal.qc.ca/}
}
\usage{
small_mtl_network
}
\description{
A SpatialLinesDataFrame representing a smaller portion of the road
network of Montreal.
}
\keyword{datasets}
