% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{snapPointsToLines2}
\alias{snapPointsToLines2}
\title{Snap points to lines}
\usage{
snapPointsToLines2(points, lines, idField = NA)
}
\arguments{
\item{points}{A SpatialPointsDataFrame}

\item{lines}{A SpatialLinesDataFrame}

\item{idField}{The name of the column to use as index for the lines}
}
\value{
A SpatialPointsDataFrame with the projected geometries
}
\description{
Snap points to their nearest lines (edited from maptools)
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
