% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indexing.R
\name{spatial_request}
\alias{spatial_request}
\title{Spatial request}
\usage{
spatial_request(geometry, tree, data)
}
\arguments{
\item{geometry}{objects such as SpatialLine, SpatialPolygon or SpatialPoint}

\item{tree}{a tree object from package SearchTrees}

\item{data}{the original data used to build the tree object}
}
\value{
a subset of data, intersecting geometry
}
\description{
Use a quadtree index to perform spatial request.
}
\examples{
#This is an internal function, no example provided
}
