% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialNetwork-methods.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,SpatialNetwork-method}
\alias{points.SpatialNetwork}
\alias{print.summary.SpatialNetwork}
\alias{summary}
\alias{summary,SpatialNetwork-method}
\alias{summary.SpatialNetwork}
\title{extract method for SpatialNetwork}
\usage{
\S4method{[}{SpatialNetwork}(x, i, j, ..., drop = TRUE)

\method{points}{SpatialNetwork}(x, ..., col = "red", cex = 2)

summary.SpatialNetwork(object, ...)

\method{print}{summary.SpatialNetwork}(x, ...)
}
\arguments{
\item{x}{object of class SpatialNetwork}

\item{i}{numeric; features to select}

\item{j}{numeric or character; attributes to select}

\item{...}{ignored}

\item{drop}{logical; ignored}

\item{col}{color}

\item{cex}{symbol size}

\item{object}{object of class SpatialNetwork}
}
\description{
extract method for SpatialNetwork

summary method for SpatialNetwork
}

