% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialNetwork-methods.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,SpatialNetwork-method}
\alias{as}
\alias{coerce,SpatialNetwork,SpatialPointsDataFrame-method}
\alias{coerce,SpatialNetwork,igraph-method}
\alias{plot}
\alias{plot,SpatialNetwork,missing-method}
\alias{points.SpatialNetwork}
\alias{print.summary.SpatialNetwork}
\alias{spTransform}
\alias{spTransform,SpatialNetwork,ANY-method}
\alias{summary}
\alias{summary,SpatialNetwork-method}
\title{extract method for SpatialNetwork}
\usage{
\S4method{[}{SpatialNetwork}(x, i, j, ..., drop = TRUE)

\method{points}{SpatialNetwork}(x, ..., col = "red", cex = 2)

\S4method{summary}{SpatialNetwork}(object, ...)

\method{print}{summary.SpatialNetwork}(x, ...)

\S4method{spTransform}{SpatialNetwork,ANY}(x, CRSobj, ...)

\S4method{plot}{SpatialNetwork,missing}(x, y, ..., arrow_size = 0)
}
\arguments{
\item{x}{object of class SpatialNetwork}

\item{i}{numeric; Lines to select (and any edges referring to them)}

\item{j}{numeric; graph edges to select (and any Lines they refer to)}

\item{...}{ignored}

\item{drop}{logical; ignored}

\item{col}{color}

\item{cex}{symbol size}

\item{object}{object of class SpatialNetwork}

\item{CRSobj}{object of class \link[sp]{CRS}}

\item{y}{ignored}

\item{arrow_size}{numeric; scaling factor for the arrow head}
}
\description{
extract method for SpatialNetwork

summary method for SpatialNetwork

spTransform method for SpatialNetwork

As("SpatialNetwork", "SpatialPointsDataFrame")

As("SpatialNetwork", "igraph")

plot method for SpatialNetwork
}
\examples{
data(torontocentre)
tc = SpatialNetwork(torontocentre, direction = torontocentre$DIRECTION_)
library(sp)
plot(tc, arrow_size = .5)
points(tc)
}

