% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border_correction.R
\name{split_border}
\alias{split_border}
\title{Split bundary of polygon}
\usage{
split_border(polygon, bw)
}
\arguments{
\item{polygon}{The polygon representing the study area}

\item{bw}{The maximum bandwidth}
}
\value{
A SpatialLinesDataFrame
}
\description{
A function to cut the boundary of the study area into chunks.
}
\keyword{internal}
