% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nkde_execution_functions_sf.R
\name{split_by_grid.mc}
\alias{split_by_grid.mc}
\title{Split data with a grid (multicore)}
\usage{
split_by_grid.mc(
  grid,
  samples,
  events,
  lines,
  bw,
  tol,
  digits,
  split_all = TRUE
)
}
\arguments{
\item{grid}{A spatial grid to split the data within}

\item{samples}{A feature collection of points representing the samples points}

\item{events}{A feature collection of points representing the events points}

\item{lines}{A feautre collection of linestrings representing the network}

\item{bw}{The kernel bandwidth (used to avoid edge effect)}

\item{tol}{A float indicating the spatial tolerance when snapping events on
lines}

\item{digits}{The number of digits to keep}

\item{split_all}{A boolean indicating if we must split the lines at each vertex
(TRUE) or only at event vertices (FALSE)}
}
\value{
A list with the split dataset
}
\description{
Function to split the dataset according to a grid with multicore support.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
