% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_functions.R
\name{split_lines_at_vertex}
\alias{split_lines_at_vertex}
\title{Split lines at vertices in a SpatialLinesDataFrame}
\usage{
split_lines_at_vertex(lines, points, nearest_lines_idx, mindist)
}
\arguments{
\item{lines}{The SpatialLinesDataframe to split}

\item{points}{The SpatialPoints to add to as vertex to the lines}

\item{nearest_lines_idx}{For each point, the index of the nearest line}

\item{mindist}{The minimum distance between one point and the extremity of
the line to add the point as a vertex.}
}
\value{
An object of class SpatialLinesDataFrame (package sp)
}
\description{
Split lines (SpatialLines) at their nearest vertices
(SpatialPoints), may fail if the lines geometries are self intersecting.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
