% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_nkde.R
\name{tnkde}
\alias{tnkde}
\title{Temporal Network Kernel density estimate}
\usage{
tnkde(
  lines,
  events,
  time_field,
  w,
  samples_loc,
  samples_time,
  kernel_name,
  bw_net,
  bw_time,
  adaptive = FALSE,
  trim_bw_net = NULL,
  trim_bw_time = NULL,
  method,
  div = "bw",
  diggle_correction = FALSE,
  study_area = NULL,
  max_depth = 15,
  digits = 5,
  tol = 0.1,
  agg = NULL,
  sparse = TRUE,
  grid_shape = c(1, 1),
  verbose = TRUE,
  check = TRUE
)
}
\arguments{
\item{lines}{A SpatialLinesDataFrame representing the underlying network. The
geometries must be a SpatialLinesDataFrame (may crash if some geometries
 are invalid) without MultiLineSring.}

\item{events}{events A SpatialPointsDataFrame representing the events on the
network. The points will be snapped on the network to their closest line.}

\item{time_field}{The name of the field in events indicating when the events occurred. It must be a numeric field}

\item{w}{A vector representing the weight of each event}

\item{samples_loc}{A SpatialPointsDataFrame representing the locations for
which the densities will be estimated.}

\item{samples_time}{A numeric vector indicating when the densities will be sampled}

\item{kernel_name}{The name of the kernel to use. Must be one of triangle,
gaussian, tricube, cosine ,triweight, quartic, epanechnikov or uniform.}

\item{bw_net}{The network kernel bandwidth (using the scale of the lines)}

\item{bw_time}{The time kernel bandwidth}

\item{adaptive}{A Boolean, indicating if an adaptive bandwidth must be
used. Both spatial and temporal banwidth are adapted but separately.}

\item{trim_bw_net}{A float, indicating the maximum value for the adaptive
network bandwidth}

\item{trim_bw_time}{A float, indicating the maximum value for the adaptive
time bandwidth}

\item{method}{The method to use when calculating the NKDE, must be one of
simple / discontinuous / continuous (see nkde details for more information)}

\item{div}{The divisor to use for the kernel. Must be "n" (the number of
events within the radius around each sampling point), "bw" (the bandwith)
"none" (the simple sum).}

\item{diggle_correction}{A Boolean indicating if the correction factor
for edge effect must be used.}

\item{study_area}{A SpatialPolygonsDataFrame or a SpatialPolygon
representing the limits of the study area.}

\item{max_depth}{when using the continuous and discontinuous methods, the
calculation time and memory use can go wild  if the network has many
small edges (area with many of intersections and many events). To
avoid it, it is possible to set here a maximum depth. Considering that the
kernel is divided at intersections, a value of 10 should yield good
estimates in most cases. A larger value can be used without a problem for the
discontinuous method. For the continuous method, a larger value will
strongly impact calculation speed.}

\item{digits}{The number of digits to retain from the spatial coordinates. It
ensures that topology is good when building the network. Default is 3. Too high a
precision (high number of digits) might break some connections}

\item{tol}{A float indicating the minimum distance between the events and the
lines' extremities when adding the point to the network. When points are
closer, they are added at the extremity of the lines.}

\item{agg}{A double indicating if the events must be aggregated within a
distance. If NULL, the events are aggregated only by rounding the
coordinates.}
}
\description{
Calculate the Temporal Network Kernel Density Estimate based on a network of lines,
sampling points in space and times, and events in space and time.
}
\details{
**Temporal Network Kernel Density Estimate**\cr
The TNKDE is an extension of the NKDE considering both the location of events on a the network and
in time. Thus, density estimation (density sampling) can be done along lines of the network and
at different time. It can be used with the three NKDE (simple, discontinuous and continuous).
\cr\cr
**density in time and space**\cr
Two bandwidths must be provided, one for the network distance and one for the time distance. They are
both used to calculate the contribution of each event to each sampling point. Let us consider one event
E and a sample S. dnet(E,S) is the network density of E at S location and dtime(E,S) is the time density of
E at S time. The total density is thus dnet(E,S) * dtime(E,S). If one of the two densities is 0, then the
tota density is 0 because the sampling point is out of the covered area by the event in time or in the
network space.
\cr\cr
**adaptive bandwidth**\cr
It is possible to use adaptive bandwidth both on the network and in time.
Adaptive bandwidths are calculated using the Abramson’s smoothing regimen \insertCite{abramson1982bandwidth}{spNetwork}.
To do so, the original fixed bandwidths must be specified (bw_net and bw_time parameters).
They are used to estimate the priory densitiy at event locations and time, separately.
These densities are then used to calculate local bandwidths. The maximum size of the two local
bandwidths can be limited with the parameters trim_bw_net and trim_bw_time.
#' \cr\cr
**Diggle correction factor**\cr
A set of events can be limited in both space (limits of the study area) and time (
begining and ending of the data collection period). These limits induce lower densities
at the border of the set of events, because they are not sampled outside the limits. It is
possible to apply the Diggle correction factor \insertCite{diggle1985kernel}{spNetwork} in both the network and time spaces to minimize
this effect.
}
