% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_functions.R
\name{tricube_kernel}
\alias{tricube_kernel}
\title{Tricube kernel}
\usage{
tricube_kernel(d, bw)
}
\arguments{
\item{d}{The distance from the event}

\item{bw}{The bandwidth used for the kernel}
}
\value{
The estimated density
}
\description{
Function implementing the tricube kernel.
}
\examples{
#This is an internal function, no example provided
}
